+++ examples/forge-sql-orm-example-ai/static/forge-orm-example/build-static/assets/body-Dxgu0KkV.js
@@ -1,5 +1,5 @@
-import { a as e } from "./rolldown-runtime-BYbx6iT9.js";
-import { n as t, r as n } from "./react-dom-vendor-BHIljUNy.js";
+import { o as e } from "./rolldown-runtime-BM3Ffeng.js";
+import { n as t, r as n } from "./react-dom-vendor-0VJF9GAl.js";
 import {
   C as r,
   D as i,
@@ -10,23 +10,23 @@ import {
   _ as l,
   a as u,
   b as d,
-  f,
+  c as f,
   g as p,
   h as m,
   i as h,
   k as g,
-  m as _,
-  n as v,
-  p as y,
-  r as b,
+  n as _,
+  o as v,
+  r as y,
+  s as b,
   t as x,
   v as S,
   w as C,
   x as w,
   y as T,
-} from "./index-CRQqC8L5.js";
+} from "./index-BrObFP80.js";
 function E(e, t, n) {
-  return ((t = a(t)), i(e, D() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
+  return ((t = T(t)), d(e, D() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
 function D() {
   try {
@@ -41,11 +41,11 @@ var O = !0,
   A = (function (e) {
     function t(e) {
       var n;
-      return (g(this, t), (n = E(this, t)), (n.message = e), n);
+      return (C(this, t), (n = E(this, t)), (n.message = e), n);
     }
     return (
-      c(t, e),
-      o(t, [
+      S(t, e),
+      r(t, [
         {
           key: `toString`,
           value: function () {
@@ -54,17 +54,17 @@ var O = !0,
         },
       ])
     );
-  })(T(Error));
+  })(l(Error));
 function j(e, t) {
   if (!e) throw O ? new A(k) : new A(`${k}: ${t || ``}`);
 }
 var M = e(n()),
-  ee = (0, M.createContext)(null);
-function N() {
-  var e = (0, M.useContext)(ee);
+  N = (0, M.createContext)(null);
+function P() {
+  var e = (0, M.useContext)(N);
   return (j(e !== null, `Unable to find DragDropContext context`), e);
 }
-function te(e) {
+function ee(e) {
   var t = e.children,
     n = e.contextId,
     r = e.getDragState,
@@ -76,13 +76,13 @@ function te(e) {
       },
       [n, r, i, a],
     );
-  return M.createElement(ee.Provider, { value: o }, t);
+  return M.createElement(N.Provider, { value: o }, t);
 }
-function P(e) {
+function F(e) {
   return { margin: window.getComputedStyle(e).margin, rect: e.getBoundingClientRect() };
 }
-function F() {
-  var e = N().getDragState,
+function te() {
+  var e = P().getDragState,
     t = e();
   return t.isDragging ? t.draggableDimensions : null;
 }
@@ -99,7 +99,7 @@ function re() {
   };
 }
 function ie() {
-  var e = r((0, M.useState)(re), 1)[0];
+  var e = g((0, M.useState)(re), 1)[0];
   return (
     (0, M.useEffect)(
       function () {
@@ -132,11 +132,11 @@ function R(e, t) {
   return (j(n !== null, `Expected '${t}' to be present`), n);
 }
 function ae(e, t) {
-  for (var n = 0, i = Object.entries(t); n < i.length; n++) {
-    var a = r(i[n], 2),
-      o = a[0],
-      s = a[1];
-    e.setAttribute(o, s);
+  for (var n = 0, r = Object.entries(t); n < r.length; n++) {
+    var i = g(r[n], 2),
+      a = i[0],
+      o = i[1];
+    e.setAttribute(a, o);
   }
   return function () {
     for (var n = 0, r = Object.keys(t); n < r.length; n++) {
@@ -203,7 +203,7 @@ function pe(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? fe(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -234,7 +234,7 @@ function ge(e, t) {
     i = t.allowedEdges,
     a = { x: r.clientX, y: r.clientY },
     o = n.getBoundingClientRect(),
-    s =
+    c =
       i
         .map(function (e) {
           return { edge: e, value: me[e](o, a) };
@@ -242,7 +242,7 @@ function ge(e, t) {
         .sort(function (e, t) {
           return e.value - t.value;
         })[0]?.edge ?? null;
-  return pe(pe({}, e), {}, C({}, he, s));
+  return pe(pe({}, e), {}, s({}, he, c));
 }
 function _e(e) {
   return e[he] ?? null;
@@ -259,9 +259,9 @@ function ye(e) {
     a = e.type;
   return (0, M.useMemo)(
     function () {
-      return C(
-        C(
-          C(C(C(C({}, ve, !0), `draggableId`, t), `droppableId`, n), `getIndex`, r),
+      return s(
+        s(
+          s(s(s(s({}, ve, !0), `draggableId`, t), `droppableId`, n), `getIndex`, r),
           `contextId`,
           i,
         ),
@@ -282,7 +282,7 @@ function Se(e) {
     r = e.getIsDropDisabled;
   return (0, M.useMemo)(
     function () {
-      return C(C(C(C({}, be, !0), `contextId`, t), `droppableId`, n), `getIsDropDisabled`, r);
+      return s(s(s(s({}, be, !0), `contextId`, t), `droppableId`, n), `getIsDropDisabled`, r);
     },
     [t, n, r],
   );
@@ -300,7 +300,7 @@ var we = [`droppableId`, `getIndex`],
 function Ee(e) {
   var t = e.droppableId,
     n = e.getIndex,
-    r = w(e, we),
+    r = c(e, we),
     i = n(),
     a = _e(r);
   return ((a === `bottom` || a === `right`) && (i += 1), { droppableId: t, index: i });
@@ -308,7 +308,7 @@ function Ee(e) {
 function De(e) {
   var t = e.contextId,
     n = e.droppableId,
-    r = w(e, Te),
+    r = c(e, Te),
     i = Ce({ contextId: t, droppableId: n });
   if (i.length === 0) return { droppableId: n, index: 0 };
   var a = _e(r);
@@ -357,7 +357,7 @@ function ke() {
   };
 }
 function Ae() {
-  return r((0, M.useState)(ke), 1)[0];
+  return g((0, M.useState)(ke), 1)[0];
 }
 function U() {
   var e = [...arguments];
@@ -372,10 +372,10 @@ function Me(e) {
   return e instanceof Element && e.hasAttribute(`data-pdnd-honey-pot`);
 }
 function Ne(e) {
-  var t = r(document.elementsFromPoint(e.x, e.y), 2),
+  var t = g(document.elementsFromPoint(e.x, e.y), 2),
     n = t[0],
-    i = t[1];
-  return n ? (Me(n) ? (i ?? null) : n) : null;
+    r = t[1];
+  return n ? (Me(n) ? (r ?? null) : n) : null;
 }
 var Pe = 2147483647,
   Fe = {
@@ -400,7 +400,7 @@ function W(e) {
 var Ie = W(function () {
     return typeof HTMLElement < `u` && typeof HTMLElement.prototype.showPopover == `function`;
   }),
-  G = f();
+  G = m();
 function Le(e, t) {
   var n = Object.keys(e);
   if (Object.getOwnPropertySymbols) {
@@ -418,7 +418,7 @@ function Re(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Le(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -648,7 +648,7 @@ function it(e) {
     pageY: e.pageY,
   };
 }
-var at = x(function (e) {
+var at = f(function (e) {
     return e();
   }),
   ot = (function () {
@@ -760,12 +760,12 @@ function ft(e) {
   function u() {
     (o.current.dropTargets.length && c({ dropTargets: [], input: o.current.input }),
       s.drop({ current: o.current, updatedSourcePayload: null }),
-      f());
+      d());
   }
-  function f() {
-    ((ct.isActive = !1), p());
+  function d() {
+    ((ct.isActive = !1), f());
   }
-  var p = (0, G.bindAll)(
+  var f = (0, G.bindAll)(
     window,
     [
       {
@@ -798,7 +798,7 @@ function ft(e) {
               current: o.current,
               updatedSourcePayload: n.type === `external` ? n.getDropPayload(e) : null,
             }),
-            f());
+            d());
         },
       },
       {
@@ -807,7 +807,7 @@ function ft(e) {
           ((o.current = { dropTargets: o.current.dropTargets, input: it(e) }), u());
         },
       },
-    ].concat(d(rt({ onDragEnd: u }))),
+    ].concat(w(rt({ onDragEnd: u }))),
     { capture: !0 },
   );
   s.start({ nativeSetDragImage: ut(t) });
@@ -869,7 +869,7 @@ function K(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? bt(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -973,19 +973,19 @@ function Tt(e) {
       return i instanceof Node ? c({ source: t, target: i.parentElement, input: o, result: l }) : l;
     var u = i.closest(a);
     if (u == null) return l;
-    var f = r.get(u);
-    if (f == null) return l;
-    var p = { input: o, source: t, element: f.element };
-    if (f.canDrop && !f.canDrop(p))
-      return c({ source: t, target: f.element.parentElement, input: o, result: l });
-    var m = f.getData?.call(f, p) ?? {},
-      h = f.getDropEffect?.call(f, p) ?? n,
-      g = { data: m, element: f.element, dropEffect: h, isActiveDueToStickiness: !1 };
+    var d = r.get(u);
+    if (d == null) return l;
+    var f = { input: o, source: t, element: d.element };
+    if (d.canDrop && !d.canDrop(f))
+      return c({ source: t, target: d.element.parentElement, input: o, result: l });
+    var p = d.getData?.call(d, f) ?? {},
+      m = d.getDropEffect?.call(d, f) ?? n,
+      h = { data: p, element: d.element, dropEffect: m, isActiveDueToStickiness: !1 };
     return c({
       source: t,
-      target: f.element.parentElement,
+      target: d.element.parentElement,
       input: o,
-      result: [].concat(d(l), [g]),
+      result: [].concat(w(l), [h]),
     });
   }
   function l(e) {
@@ -1061,10 +1061,10 @@ function Tt(e) {
       }
     },
   };
-  function f(e) {
+  function d(e) {
     u[e.eventName](e);
   }
-  function p(e) {
+  function f(e) {
     var t = e.source,
       n = e.target,
       i = e.input,
@@ -1090,7 +1090,7 @@ function Tt(e) {
     }
     return wt(u);
   }
-  return { dropTargetForConsumers: s, getIsOver: p, dispatchEvent: f };
+  return { dropTargetForConsumers: s, getIsOver: f, dispatchEvent: d };
 }
 function Et(e, t) {
   var n = (typeof Symbol < `u` && e[Symbol.iterator]) || e[`@@iterator`];
@@ -1172,7 +1172,7 @@ function At(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? kt(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1279,44 +1279,44 @@ var zt = Ye(),
             if (e.canStart(t) && !t.defaultPrevented && t.dataTransfer) {
               var n = t.target;
               if (n instanceof HTMLElement) {
-                var i = Lt.get(n);
-                if (i) {
-                  var a = it(t),
-                    o = { element: i.element, dragHandle: i.dragHandle ?? null, input: a };
-                  if (i.canDrag && !i.canDrag(o)) {
+                var r = Lt.get(n);
+                if (r) {
+                  var i = it(t),
+                    a = { element: r.element, dragHandle: r.dragHandle ?? null, input: i };
+                  if (r.canDrag && !r.canDrag(a)) {
                     t.preventDefault();
                     return;
                   }
-                  if (i.dragHandle) {
-                    var s = Ne({ x: a.clientX, y: a.clientY });
-                    if (!i.dragHandle.contains(s)) {
+                  if (r.dragHandle) {
+                    var o = Ne({ x: i.clientX, y: i.clientY });
+                    if (!r.dragHandle.contains(o)) {
                       t.preventDefault();
                       return;
                     }
                   }
-                  var c = i.getInitialDataForExternal?.call(i, o) ?? null;
-                  if (c)
-                    for (var l = 0, u = Object.entries(c); l < u.length; l++) {
-                      var d = r(u[l], 2),
-                        f = d[0],
-                        p = d[1];
-                      t.dataTransfer.setData(f, p ?? ``);
+                  var s = r.getInitialDataForExternal?.call(r, a) ?? null;
+                  if (s)
+                    for (var c = 0, l = Object.entries(s); c < l.length; c++) {
+                      var u = g(l[c], 2),
+                        d = u[0],
+                        f = u[1];
+                      t.dataTransfer.setData(d, f ?? ``);
                     }
                   (Nt() &&
                     !t.dataTransfer.types.includes(`text/plain`) &&
                     !t.dataTransfer.types.includes(`text/uri-list`) &&
                     t.dataTransfer.setData(Ft, Pt),
                     t.dataTransfer.setData(It, ``));
-                  var m = {
+                  var p = {
                     type: `element`,
                     payload: {
-                      element: i.element,
-                      dragHandle: i.dragHandle ?? null,
-                      data: i.getInitialData?.call(i, o) ?? {},
+                      element: r.element,
+                      dragHandle: r.dragHandle ?? null,
+                      data: r.getInitialData?.call(r, a) ?? {},
                     },
                     startedFrom: `internal`,
                   };
-                  e.start({ event: t, dragType: m });
+                  e.start({ event: t, dragType: p });
                 }
               }
             }
@@ -1339,7 +1339,7 @@ function Ut(e) {
   return W(U(Bt.registerUsage(), Rt(e), yt(e.element, { attribute: `draggable`, value: `true` })));
 }
 function Wt(e, t, n) {
-  return ((t = a(t)), i(e, Gt() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
+  return ((t = T(t)), d(e, Gt() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
 function Gt() {
   try {
@@ -1385,11 +1385,11 @@ function Kt(e) {
 }
 var qt = (function (e) {
   function t() {
-    return (g(this, t), Wt(this, t, arguments));
+    return (C(this, t), Wt(this, t, arguments));
   }
   return (
-    c(t, e),
-    o(
+    S(t, e),
+    r(
       t,
       [
         {
@@ -1430,7 +1430,7 @@ function Yt(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Jt(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1534,7 +1534,7 @@ function q(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? rn(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1679,12 +1679,12 @@ function ln(e) {
           mode: `SNAP`,
         });
         var u = c(),
-          f = R(B({ attribute: I.droppable.id, value: u.droppableId }), L.droppable.direction);
-        j(f === `vertical` || f === `horizontal`);
-        function p() {
+          d = R(B({ attribute: I.droppable.id, value: u.droppableId }), L.droppable.direction);
+        j(d === `vertical` || d === `horizontal`);
+        function f() {
           t.stopDrag({ reason: `CANCEL` });
         }
-        var m = [
+        var p = [
           `mousedown`,
           `mouseup`,
           `click`,
@@ -1693,7 +1693,7 @@ function ln(e) {
           `wheel`,
           `visibilitychange`,
         ].map(function (e) {
-          return { type: e, listener: p };
+          return { type: e, listener: f };
         });
         i(
           (0, G.bindAll)(
@@ -1712,12 +1712,12 @@ function ln(e) {
                       (e.preventDefault(), t.stopDrag({ reason: `CANCEL` }));
                       return;
                     }
-                    (i = (o = cn[f])[e.key]) == null ||
+                    (i = (o = cn[d])[e.key]) == null ||
                       i.call(o, e, { dragController: t, droppableRegistry: n, contextId: r });
                   }
                 },
               },
-            ].concat(d(m)),
+            ].concat(w(p)),
           ),
         );
       },
@@ -2087,7 +2087,7 @@ function Qn() {
           },
         },
         { type: `dragend`, listener: $n },
-      ].concat(d(rt({ onDragEnd: $n }))),
+      ].concat(w(rt({ onDragEnd: $n }))),
       { capture: !1 },
     )));
 }
@@ -2167,7 +2167,7 @@ function ar(e) {
     n = e.styles;
   return `${t} { ${Object.entries(n)
     .map(function (e) {
-      var t = r(e, 2);
+      var t = g(e, 2);
       return `${t[0]}: ${t[1]};`;
     })
     .join(` `)} }`;
@@ -2320,26 +2320,26 @@ function gr() {
   };
 }
 function _r() {
-  return r((0, M.useState)(gr), 1)[0];
+  return g((0, M.useState)(gr), 1)[0];
 }
 var vr = (0, M.createContext)(null);
 function yr(e) {
   var t = e.children,
     n = e.lifecycle,
-    i = (0, M.useCallback)(
+    r = (0, M.useCallback)(
       function (e) {
-        for (var t = [], i = 0, a = Object.entries(e); i < a.length; i++) {
-          var o = a[i],
-            s = r(o, 2),
-            c = s[0],
-            l = s[1];
-          t.push(n.addResponder(c, l));
+        for (var t = [], r = 0, i = Object.entries(e); r < i.length; r++) {
+          var a = i[r],
+            o = g(a, 2),
+            s = o[0],
+            c = o[1];
+          t.push(n.addResponder(s, c));
         }
         return U.apply(void 0, t);
       },
       [n],
     );
-  return M.createElement(vr.Provider, { value: i }, t);
+  return M.createElement(vr.Provider, { value: r }, t);
 }
 function br() {
   var e = (0, M.useContext)(vr);
@@ -2445,7 +2445,7 @@ function Mr() {
   return { schedule: t, flush: n };
 }
 function Nr() {
-  return r((0, M.useState)(Mr), 1)[0];
+  return g((0, M.useState)(Mr), 1)[0];
 }
 var Pr = 0;
 function Fr() {
@@ -2458,22 +2458,22 @@ function Ir(e) {
 function Lr(e) {
   var t = e.children,
     n = e.dragHandleUsageInstructions,
-    i = n === void 0 ? Ar : n,
-    a = e.nonce,
-    o = e.onBeforeCapture,
-    s = e.onBeforeDragStart,
-    c = e.onDragStart,
-    l = e.onDragUpdate,
-    u = e.onDragEnd,
-    d = r((0, M.useState)(Fr), 1)[0];
-  Qt({ contextId: d, text: i });
-  var f = _r(),
-    p = Nr(),
-    m = p.schedule,
-    h = p.flush,
-    g = (0, M.useRef)({ isDragging: !1 }),
+    r = n === void 0 ? Ar : n,
+    i = e.nonce,
+    a = e.onBeforeCapture,
+    o = e.onBeforeDragStart,
+    s = e.onDragStart,
+    c = e.onDragUpdate,
+    l = e.onDragEnd,
+    u = g((0, M.useState)(Fr), 1)[0];
+  Qt({ contextId: u, text: r });
+  var d = _r(),
+    f = Nr(),
+    p = f.schedule,
+    m = f.flush,
+    h = (0, M.useRef)({ isDragging: !1 }),
     _ = (0, M.useCallback)(function () {
-      return g.current;
+      return h.current;
     }, []),
     v = Ae(),
     y = (0, M.useCallback)(
@@ -2502,35 +2502,35 @@ function Lr(e) {
         var t = e.targetLocation,
           n = e.isImmediate,
           r = n === void 0 ? !1 : n;
-        if (!g.current.isDragging) return;
-        var i = g.current,
+        if (!h.current.isDragging) return;
+        var i = h.current,
           a = i.prevDestination,
           o = i.draggableId,
           s = i.type,
-          c = i.sourceLocation,
-          u = Xt({ start: c, target: t });
+          l = i.sourceLocation,
+          u = Xt({ start: l, target: t });
         if (H(a, u)) return;
-        Object.assign(g.current, { prevDestination: u, sourceLocation: c, targetLocation: t });
-        var d = {
-            mode: g.current.mode,
+        Object.assign(h.current, { prevDestination: u, sourceLocation: l, targetLocation: t });
+        var f = {
+            mode: h.current.mode,
             draggableId: o,
             type: s,
-            source: c,
+            source: l,
             destination: u,
             combine: null,
           },
-          p = t ? v.getEntry({ droppableId: t.droppableId }) : null;
-        (f.dispatch(`onPrePendingDragUpdate`, { update: d, targetLocation: t }),
-          f.dispatch(`onPendingDragUpdate`, { update: d, targetLocation: t, droppable: p }));
-        function h() {
-          var e = kr(`onDragUpdate`, d),
+          m = t ? v.getEntry({ droppableId: t.droppableId }) : null;
+        (d.dispatch(`onPrePendingDragUpdate`, { update: f, targetLocation: t }),
+          d.dispatch(`onPendingDragUpdate`, { update: f, targetLocation: t, droppable: m }));
+        function g() {
+          var e = kr(`onDragUpdate`, f),
             t = e.provided,
             n = e.getMessage;
-          (l?.(d, t), Tr(n()));
+          (c?.(f, t), Tr(n()));
         }
-        r ? h() : m(h);
+        r ? g() : p(g);
       },
-      [v, f, l, m],
+      [v, d, c, p],
     ),
     x = (0, M.useCallback)(
       function (e) {
@@ -2538,40 +2538,40 @@ function Lr(e) {
           n = e.type,
           r = e.getSourceLocation,
           i = e.sourceElement,
-          a = e.mode;
-        if (!g.current.isDragging) {
-          o?.({ draggableId: t, mode: a });
-          var l = { mode: a, draggableId: t, type: n, source: r() },
+          c = e.mode;
+        if (!h.current.isDragging) {
+          a?.({ draggableId: t, mode: c });
+          var l = { mode: c, draggableId: t, type: n, source: r() },
             u = document.activeElement,
-            d =
+            f =
               u instanceof HTMLElement && u.hasAttribute(I.dragHandle.draggableId)
                 ? R(u, I.dragHandle.draggableId)
                 : null,
-            p = l.source.droppableId,
-            h = v.getEntry({ droppableId: p });
-          (j(h, `should have entry for droppable '${p}'`),
-            (g.current = {
+            m = l.source.droppableId,
+            g = v.getEntry({ droppableId: m });
+          (j(g, `should have entry for droppable '${m}'`),
+            (h.current = {
               isDragging: !0,
-              mode: a,
-              draggableDimensions: P(i),
-              restoreFocusTo: d,
+              mode: c,
+              draggableDimensions: F(i),
+              restoreFocusTo: f,
               draggableId: t,
               type: n,
               prevDestination: l.source,
               sourceLocation: l.source,
               targetLocation: l.source,
-              draggableInitialOffsetInSourceDroppable: Ir({ element: i, mode: h.mode }),
+              draggableInitialOffsetInSourceDroppable: Ir({ element: i, mode: g.mode }),
             }),
-            s?.(l),
-            f.dispatch(`onPendingDragStart`, { start: l, droppable: h }),
-            m(function () {
-              var e = { mode: a, draggableId: t, type: n, source: r() },
+            o?.(l),
+            d.dispatch(`onPendingDragStart`, { start: l, droppable: g }),
+            p(function () {
+              var e = { mode: c, draggableId: t, type: n, source: r() },
                 i = kr(`onDragStart`, e),
-                o = i.provided,
-                s = i.getMessage;
-              (c?.(e, o),
-                Tr(s()),
-                m(function () {
+                a = i.provided,
+                o = i.getMessage;
+              (s?.(e, a),
+                Tr(o()),
+                p(function () {
                   var t = e.source.droppableId,
                     n = v.getEntry({ droppableId: t });
                   n != null &&
@@ -2581,46 +2581,46 @@ function Lr(e) {
             }));
         }
       },
-      [v, y, f, o, s, c, m, b],
+      [v, y, d, a, o, s, p, b],
     ),
     S = ie(),
     C = (0, M.useCallback)(
       function (e) {
         var t = e.reason;
-        if (g.current.isDragging) {
+        if (h.current.isDragging) {
           (S.runCleanupFn(), t === `CANCEL` && b({ targetLocation: null }));
-          var n = g.current,
+          var n = h.current,
             r = n.mode,
             i = n.restoreFocusTo,
             a = n.sourceLocation,
             o = n.targetLocation,
             s = n.type,
             c = n.draggableId;
-          ((g.current = { isDragging: !1 }), h());
-          var l = Xt({ start: a, target: o }),
+          ((h.current = { isDragging: !1 }), m());
+          var f = Xt({ start: a, target: o }),
             p = {
-              reason: l === null ? `CANCEL` : `DROP`,
+              reason: f === null ? `CANCEL` : `DROP`,
               type: s,
               source: a,
-              destination: l,
+              destination: f,
               mode: r,
               draggableId: c,
               combine: null,
             };
-          f.dispatch(`onBeforeDragEnd`, { draggableId: c });
-          var m = kr(`onDragEnd`, p),
-            _ = m.provided,
-            v = m.getMessage;
-          (u(p, _),
+          d.dispatch(`onBeforeDragEnd`, { draggableId: c });
+          var g = kr(`onDragEnd`, p),
+            _ = g.provided,
+            v = g.getMessage;
+          (l(p, _),
             Tr(v()),
             i &&
               requestAnimationFrame(function () {
-                var e = ce({ contextId: d, draggableId: c });
+                var e = ce({ contextId: u, draggableId: c });
                 e && e.focus();
               }));
         }
       },
-      [d, h, S, f, u, b],
+      [u, m, S, d, l, b],
     ),
     w = (0, M.useMemo)(
       function () {
@@ -2628,16 +2628,16 @@ function Lr(e) {
       },
       [_, x, C, b],
     );
-  nr({ dragController: w, contextId: d });
+  nr({ dragController: w, contextId: u });
   var T = ln({
       dragController: w,
       droppableRegistry: v,
-      contextId: d,
+      contextId: u,
       setKeyboardCleanupFn: S.setCleanupFn,
     }).startKeyboardDrag,
     E = (0, M.useCallback)(
       function (e) {
-        var t = g.current;
+        var t = h.current;
         t.isDragging &&
           e.isDropDisabled &&
           e.droppableId === t.targetLocation?.droppableId &&
@@ -2647,16 +2647,16 @@ function Lr(e) {
     );
   return (
     v.setUpdateListener(E),
-    lr({ contextId: d, nonce: a }),
+    lr({ contextId: u, nonce: i }),
     M.createElement(
       qt,
-      { contextId: d, dragController: w },
+      { contextId: u, dragController: w },
       M.createElement(
         yr,
-        { lifecycle: f },
+        { lifecycle: d },
         M.createElement(
-          te,
-          { contextId: d, getDragState: _, startKeyboardDrag: T, droppableRegistry: v },
+          ee,
+          { contextId: u, getDragState: _, startKeyboardDrag: T, droppableRegistry: v },
           t,
         ),
       ),
@@ -2721,7 +2721,7 @@ function Wr(e) {
   );
 }
 function Gr() {
-  return { startKeyboardDrag: N().startKeyboardDrag };
+  return { startKeyboardDrag: P().startKeyboardDrag };
 }
 function Kr() {
   return z({ attribute: L.dropIndicator });
@@ -2761,7 +2761,7 @@ function Zr(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Xr(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2839,7 +2839,7 @@ function oi(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ai(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2852,10 +2852,10 @@ function oi(e) {
 var si = (0, M.memo)(
     (0, M.forwardRef)(function (e, t) {
       var n = e.style,
-        r = F(),
-        i = N().contextId,
-        a = C({}, I.placeholder.contextId, i),
-        o = (0, M.useMemo)(
+        r = te(),
+        i = P().contextId,
+        o = s({}, I.placeholder.contextId, i),
+        c = (0, M.useMemo)(
           function () {
             if (r) {
               var e = r.margin,
@@ -2868,7 +2868,7 @@ var si = (0, M.memo)(
           },
           [r, n],
         );
-      return M.createElement(`div`, s({ ref: t, style: o }, a));
+      return M.createElement(`div`, a({ ref: t, style: c }, o));
     }),
   ),
   Q = {
@@ -2924,7 +2924,7 @@ function $(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ci(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2943,7 +2943,7 @@ var di = {
     var t = Q[e.direction],
       n = t.mainAxis,
       r = t.crossAxis;
-    return C(C({}, n.name, 0), r.name, 24);
+    return s(s({}, n.name, 0), r.name, 24);
   },
   home: function (e) {
     var t = e.droppable.direction,
@@ -2953,8 +2953,8 @@ var di = {
     var i = { x: n.x - r.rect.x, y: n.y - r.rect.y },
       a = Q[t],
       o = a.mainAxis,
-      s = a.crossAxis;
-    return C(C({}, o.name, i[o.name]), s.name, i[s.name] + 24);
+      c = a.crossAxis;
+    return s(s({}, o.name, i[o.name]), c.name, i[c.name] + 24);
   },
   away: function (e) {
     var t = e.droppable.direction,
@@ -2964,8 +2964,8 @@ var di = {
     var i = { x: n.x - r.rect.x, y: n.y - r.rect.y },
       a = Q[t],
       o = a.mainAxis,
-      s = a.crossAxis;
-    return C(C({}, o.name, i[o.name] - 0.5 * r.rect[o.style.length]), s.name, i[s.name] + 24);
+      c = a.crossAxis;
+    return s(s({}, o.name, i[o.name] - 0.5 * r.rect[o.style.length]), c.name, i[c.name] + 24);
   },
 };
 function fi(e, t) {
@@ -3067,53 +3067,53 @@ var gi = function () {};
 function _i(e) {
   var t = e.children,
     n = e.draggableId,
-    i = e.index,
-    a = e.isDragDisabled,
-    o = a === void 0 ? !1 : a,
-    s = e.disableInteractiveElementBlocking,
-    c = s === void 0 ? !1 : s,
+    r = e.index,
+    i = e.isDragDisabled,
+    a = i === void 0 ? !1 : i,
+    o = e.disableInteractiveElementBlocking,
+    c = o === void 0 ? !1 : o,
     l = Hr(),
     u = l.direction,
     d = l.droppableId,
     f = l.type,
-    p = l.mode,
-    m = N(),
-    h = m.contextId,
-    g = m.getDragState,
-    _ = (0, M.useRef)(null),
-    v = (0, M.useRef)(null),
-    b = ie(),
-    x = b.setCleanupFn,
-    S = b.runCleanupFn,
+    m = l.mode,
+    h = P(),
+    _ = h.contextId,
+    v = h.getDragState,
+    y = (0, M.useRef)(null),
+    b = (0, M.useRef)(null),
+    x = ie(),
+    S = x.setCleanupFn,
+    C = x.runCleanupFn,
     w = (0, M.useCallback)(
       function (e) {
-        (_.current && S(),
-          e && x(ae(e, C(C({}, L.draggable.droppableId, d), L.draggable.index, String(i)))),
-          (_.current = e),
-          (v.current = ce({ contextId: h, draggableId: n })));
+        (y.current && C(),
+          e && S(ae(e, s(s({}, L.draggable.droppableId, d), L.draggable.index, String(r)))),
+          (y.current = e),
+          (b.current = ce({ contextId: _, draggableId: n })));
       },
-      [h, n, d, i, S, x],
+      [_, n, d, r, C, S],
     ),
-    T = Jr(i),
-    E = r((0, M.useReducer)(mi, li), 2),
+    T = Jr(r),
+    E = g((0, M.useReducer)(mi, li), 2),
     D = E[0],
     O = E[1],
-    k = ye({ draggableId: n, droppableId: d, getIndex: T, contextId: h, type: f }),
+    k = ye({ draggableId: n, droppableId: d, getIndex: T, contextId: _, type: f }),
     A = D.type === `dragging`,
-    ee = D.type === `hiding`,
-    te = Hr(),
-    P = te.shouldRenderCloneWhileDragging,
-    ne = te.isDropDisabled,
+    N = D.type === `hiding`,
+    ee = Hr(),
+    F = ee.shouldRenderCloneWhileDragging,
+    ne = ee.isDropDisabled,
     re = br(),
     R = Gr().startKeyboardDrag;
   ((0, M.useEffect)(
     function () {
-      if (D.type === `idle` && !o) {
-        var e = _.current;
-        y(e instanceof HTMLElement);
-        var t = v.current;
+      if (D.type === `idle` && !a) {
+        var e = y.current;
+        p(e instanceof HTMLElement);
+        var t = b.current;
         return (
-          y(t instanceof HTMLElement),
+          p(t instanceof HTMLElement),
           (0, G.bindAll)(t, [
             {
               type: `keydown`,
@@ -3137,14 +3137,14 @@ function _i(e) {
         );
       }
     },
-    [c, n, d, T, o, R, D.type, f],
+    [c, n, d, T, a, R, D.type, f],
   ),
     (0, M.useEffect)(
       function () {
-        if (!ee && !o) {
-          var e = _.current;
+        if (!N && !a) {
+          var e = y.current;
           j(e instanceof HTMLElement);
-          var t = v.current;
+          var t = b.current;
           return (
             j(t instanceof HTMLElement),
             Ut({
@@ -3166,11 +3166,11 @@ function _i(e) {
           );
         }
       },
-      [k, c, o, A, ee],
+      [k, c, a, A, N],
     ));
-  var oe = D.type !== `idle` && p === `standard`,
+  var oe = D.type !== `idle` && m === `standard`,
     z = (0, M.useRef)(null);
-  Wr({ elementRef: oe ? z : _, data: k, direction: u, contextId: h, isDropDisabled: ne, type: f });
+  Wr({ elementRef: oe ? z : y, data: k, direction: u, contextId: _, isDropDisabled: ne, type: f });
   var se = (0, M.useRef)(!0);
   ((0, M.useEffect)(function () {
     return (
@@ -3182,18 +3182,18 @@ function _i(e) {
   }, []),
     (0, M.useEffect)(
       function () {
-        var e = g();
-        P &&
+        var e = v();
+        F &&
           e.isDragging &&
           e.draggableId === k.draggableId &&
           O({ type: `START_HIDING`, payload: { mode: e.mode } });
       },
-      [k.draggableId, g, P],
+      [k.draggableId, v, F],
     ));
-  var B = F();
+  var B = te();
   (0, M.useEffect)(
     function () {
-      return P
+      return F
         ? re({
             onPendingDragStart: function (e) {
               var t = e.start;
@@ -3213,7 +3213,7 @@ function _i(e) {
                   if (t.mode === `FLUID`)
                     return O({ type: `START_POINTER_DRAG`, payload: { start: t } });
                   if (t.mode === `SNAP`) {
-                    var r = g();
+                    var r = v();
                     return (
                       j(r.isDragging && r.draggableDimensions),
                       O({
@@ -3235,8 +3235,8 @@ function _i(e) {
                   (O({ type: `UPDATE_DRAG`, payload: { update: t } }),
                   t.mode === `SNAP` &&
                     queueMicrotask(function () {
-                      if (g().isDragging) {
-                        var e = qr(h),
+                      if (v().isDragging) {
+                        var e = qr(_),
                           r = e ? e.getBoundingClientRect() : null,
                           i = Kr();
                         O({
@@ -3271,21 +3271,21 @@ function _i(e) {
             }),
           );
     },
-    [k.draggableId, k.contextId, re, P, u, h, B, g],
+    [k.draggableId, k.contextId, re, F, u, _, B, v],
   );
   var le = (0, M.useMemo)(
       function () {
         return {
-          draggableProps: C(
-            C(C({}, I.draggable.contextId, h), I.draggable.id, n),
+          draggableProps: s(
+            s(s({}, I.draggable.contextId, _), I.draggable.id, n),
             `style`,
             ti({ draggableDimensions: B, draggableState: D }),
           ),
-          dragHandleProps: C(
-            C(
-              C(
-                C(
-                  C({ role: `button`, "aria-describedby": Zt(h) }, I.dragHandle.contextId, h),
+          dragHandleProps: s(
+            s(
+              s(
+                s(
+                  s({ role: `button`, "aria-describedby": Zt(_) }, I.dragHandle.contextId, _),
                   I.dragHandle.draggableId,
                   n,
                 ),
@@ -3301,19 +3301,19 @@ function _i(e) {
           innerRef: w,
         };
       },
-      [h, n, B, D, w],
+      [_, n, B, D, w],
     ),
     ue = hi({ draggingOver: D.draggingOver, isClone: !1, isDragging: A, mode: A ? D.mode : null }),
     de = (0, M.useMemo)(
       function () {
-        return { draggableId: n, type: f, source: { droppableId: d, index: i } };
+        return { draggableId: n, type: f, source: { droppableId: d, index: r } };
       },
-      [n, d, i, f],
+      [n, d, r, f],
     );
   return M.createElement(
     M.Fragment,
     null,
-    ee ? null : t(le, ue, de),
+    N ? null : t(le, ue, de),
     oe && M.createElement(si, { ref: z }),
   );
 }
@@ -3327,29 +3327,29 @@ function yi(e) {
     i = e.draggableId,
     a = e.index,
     o = e.draggingOver,
-    s = e.style,
-    c = e.getContainerForClone,
-    l = c === void 0 ? vi : c,
-    u = e.mode,
-    d = N().contextId,
-    f = (0, M.useCallback)(
+    c = e.style,
+    l = e.getContainerForClone,
+    u = l === void 0 ? vi : l,
+    d = e.mode,
+    f = P().contextId,
+    p = (0, M.useCallback)(
       function (e) {
-        e && ce({ contextId: d, draggableId: i })?.focus();
+        e && ce({ contextId: f, draggableId: i })?.focus();
       },
-      [d, i],
+      [f, i],
     );
   return (0, X.createPortal)(
     t(
       (0, M.useMemo)(
         function () {
           return {
-            innerRef: f,
-            draggableProps: C(C(C({}, I.draggable.contextId, d), I.draggable.id, i), `style`, s),
-            dragHandleProps: C(
-              C(
-                C(
-                  C(
-                    C({ role: `button`, "aria-describedby": Zt(d) }, I.dragHandle.contextId, d),
+            innerRef: p,
+            draggableProps: s(s(s({}, I.draggable.contextId, f), I.draggable.id, i), `style`, c),
+            dragHandleProps: s(
+              s(
+                s(
+                  s(
+                    s({ role: `button`, "aria-describedby": Zt(f) }, I.dragHandle.contextId, f),
                     I.dragHandle.draggableId,
                     i,
                   ),
@@ -3364,9 +3364,9 @@ function yi(e) {
             ),
           };
         },
-        [d, i, f, s],
+        [f, i, p, c],
       ),
-      hi({ draggingOver: o, isClone: !0, isDragging: !0, mode: u }),
+      hi({ draggingOver: o, isClone: !0, isDragging: !0, mode: d }),
       (0, M.useMemo)(
         function () {
           return { draggableId: i, type: r, source: { droppableId: n, index: a } };
@@ -3374,38 +3374,38 @@ function yi(e) {
         [i, n, a, r],
       ),
     ),
-    l(),
+    u(),
   );
 }
 function bi(e) {
   var t = e.children,
     n = e.droppableId,
-    i = e.type,
-    a = e.getContainerForClone,
-    o = N(),
-    s = o.contextId,
-    c = o.getDragState,
-    l = F(),
-    u = r((0, M.useReducer)(mi, li), 2),
-    d = u[0],
-    f = u[1],
-    p = br();
+    r = e.type,
+    i = e.getContainerForClone,
+    a = P(),
+    o = a.contextId,
+    s = a.getDragState,
+    c = te(),
+    l = g((0, M.useReducer)(mi, li), 2),
+    u = l[0],
+    d = l[1],
+    f = br();
   if (
     ((0, M.useEffect)(
       function () {
         return U(
-          p({
+          f({
             onPendingDragStart: function (e) {
               var t = e.start,
                 r = e.droppable;
               if (n === t.source.droppableId) {
                 if (t.mode === `FLUID`)
-                  return f({ type: `START_POINTER_DRAG`, payload: { start: t } });
+                  return d({ type: `START_POINTER_DRAG`, payload: { start: t } });
                 if (t.mode === `SNAP`) {
-                  var i = c();
+                  var i = s();
                   return (
                     j(i.isDragging && i.draggableDimensions),
-                    f({
+                    d({
                       type: `START_KEYBOARD_DRAG`,
                       payload: {
                         start: t,
@@ -3420,20 +3420,20 @@ function bi(e) {
             onPendingDragUpdate: function (e) {
               var t = e.update,
                 n = e.droppable;
-              d.type === `dragging` &&
-                d.draggableId === t.draggableId &&
-                (f({ type: `UPDATE_DRAG`, payload: { update: t } }),
+              u.type === `dragging` &&
+                u.draggableId === t.draggableId &&
+                (d({ type: `UPDATE_DRAG`, payload: { update: t } }),
                 t.mode === `SNAP` &&
                   queueMicrotask(function () {
-                    if (c().isDragging) {
-                      var e = qr(s),
+                    if (s().isDragging) {
+                      var e = qr(o),
                         r = e ? e.getBoundingClientRect() : null,
                         i = Kr();
-                      f({
+                      d({
                         type: `UPDATE_KEYBOARD_PREVIEW`,
                         payload: {
                           update: t,
-                          draggableDimensions: l,
+                          draggableDimensions: c,
                           droppable: n,
                           placeholderRect: r,
                           dropIndicatorRect: i ? i.getBoundingClientRect() : null,
@@ -3444,38 +3444,38 @@ function bi(e) {
             },
             onBeforeDragEnd: function (e) {
               var t = e.draggableId;
-              d.type === `dragging` && t === d.draggableId && f({ type: `DROP` });
+              u.type === `dragging` && t === u.draggableId && d({ type: `DROP` });
             },
           }),
           Ht({
             canMonitor: function (e) {
               var t = e.source;
-              return V(t.data) ? t.data.contextId === s && t.data.droppableId === n : !1;
+              return V(t.data) ? t.data.contextId === o && t.data.droppableId === n : !1;
             },
             onDrag: function (e) {
               var t = e.location;
-              f({ type: `UPDATE_POINTER_PREVIEW`, payload: { pointerLocation: t } });
+              d({ type: `UPDATE_POINTER_PREVIEW`, payload: { pointerLocation: t } });
             },
           }),
         );
       },
-      [n, s, p, d, l, c],
+      [n, o, f, u, c, s],
     ),
-    d.type !== `dragging`)
+    u.type !== `dragging`)
   )
     return null;
-  var m = ti({ draggableDimensions: l, draggableState: d });
+  var p = ti({ draggableDimensions: c, draggableState: u });
   return M.createElement(
     yi,
     {
       droppableId: n,
-      type: i,
-      draggableId: d.draggableId,
-      index: d.start.index,
-      draggingOver: d.draggingOver,
-      mode: d.mode,
-      style: m,
-      getContainerForClone: a,
+      type: r,
+      draggableId: u.draggableId,
+      index: u.start.index,
+      draggingOver: u.draggingOver,
+      mode: u.mode,
+      style: p,
+      getContainerForClone: i,
     },
     t,
   );
@@ -3587,16 +3587,16 @@ function ji(e) {
   return c ? Di({ element: c, isForwardEdge: !0, mode: i, direction: r, contextId: a }) : null;
 }
 var Mi = 4,
-  Ni = h({
+  Ni = v({
     background: `var(--ds-border-brand, #1868DB)`,
     scrollMarginTop: Mi,
     scrollMarginBottom: Mi + 1,
   }),
-  Pi = h({ position: `absolute`, top: 0, left: 0 }),
-  Fi = h({ opacity: 0 }),
+  Pi = v({ position: `absolute`, top: 0, left: 0 }),
+  Fi = v({ opacity: 0 }),
   Ii = {
-    horizontal: h({ width: 2, height: `100%`, marginLeft: -2 }),
-    vertical: h({ width: `100%`, height: 2, marginTop: -2 }),
+    horizontal: v({ width: 2, height: `100%`, marginLeft: -2 }),
+    vertical: v({ width: `100%`, height: 2, marginTop: -2 }),
   };
 function Li(e) {
   var t = e.direction,
@@ -3606,8 +3606,8 @@ function Li(e) {
   var i = Q[t],
     a = i.mainAxis,
     o = i.crossAxis;
-  return C(
-    C(
+  return s(
+    s(
       { transform: `${a.style.transform}(${n.mainAxis.offset - r}px)` },
       o.style.length,
       n.crossAxis.length,
@@ -3616,34 +3616,34 @@ function Li(e) {
     n.crossAxis.offset,
   );
 }
-var Ri = C({}, L.dropIndicator, ``),
+var Ri = s({}, L.dropIndicator, ``),
   zi = function (e) {
     var t = e.direction,
       n = e.mode,
-      i = N(),
-      a = i.contextId,
-      o = i.getDragState,
-      c = (0, M.useRef)(null),
-      l = r((0, M.useState)(null), 2),
-      d = l[0],
-      f = l[1],
-      p = r((0, M.useState)(!1), 2),
-      m = p[0],
-      h = p[1],
-      g = br(),
-      _ = (0, M.useCallback)(
+      r = P(),
+      i = r.contextId,
+      o = r.getDragState,
+      s = (0, M.useRef)(null),
+      c = g((0, M.useState)(null), 2),
+      l = c[0],
+      u = c[1],
+      d = g((0, M.useState)(!1), 2),
+      f = d[0],
+      p = d[1],
+      m = br(),
+      h = (0, M.useCallback)(
         function (e) {
           var r = e.targetLocation,
-            i = e.source,
+            a = e.source,
             o = e.destination;
-          if (!r) return f(null);
-          var s = H(i, o);
+          if (!r) return u(null);
+          var s = H(a, o);
           return (
-            h(s),
-            f(ji({ targetLocation: r, isInHomeLocation: s, direction: t, mode: n, contextId: a }))
+            p(s),
+            u(ji({ targetLocation: r, isInHomeLocation: s, direction: t, mode: n, contextId: i }))
           );
         },
-        [a, t, n],
+        [i, t, n],
       );
     ((0, M.useLayoutEffect)(
       function () {
@@ -3652,38 +3652,38 @@ var Ri = C({}, L.dropIndicator, ``),
           var t = e.targetLocation,
             n = e.sourceLocation;
           return (
-            _({ targetLocation: t, destination: Xt({ start: n, target: t }), source: n }),
-            g({
+            h({ targetLocation: t, destination: Xt({ start: n, target: t }), source: n }),
+            m({
               onPrePendingDragUpdate: function (e) {
                 var t = e.update,
                   n = e.targetLocation,
                   r = t.destination,
                   i = r === void 0 ? null : r,
                   a = t.source;
-                _({ targetLocation: n, source: a, destination: i });
+                h({ targetLocation: n, source: a, destination: i });
               },
             })
           );
         }
       },
-      [a, t, o, n, g, _],
+      [i, t, o, n, m, h],
     ),
       (0, M.useLayoutEffect)(
         function () {
-          if (d !== null) {
+          if (l !== null) {
             var e = o();
             if (!(!e.isDragging || e.mode !== `SNAP`)) {
-              var t = c.current;
+              var t = s.current;
               (j(t instanceof HTMLElement), t.scrollIntoView({ block: `nearest` }));
             }
           }
         },
-        [d, o],
+        [l, o],
       ));
-    var v = Q[t].mainAxis,
-      y = Li({ direction: t, dimensions: d, indicatorOffset: c.current ? c.current[v.offset] : 0 }),
-      b = n === `virtual`;
-    return u(`div`, s({ ref: c, css: [Ni, Ii[t], b && Pi, m && Fi], style: y }, Ri));
+    var _ = Q[t].mainAxis,
+      v = Li({ direction: t, dimensions: l, indicatorOffset: s.current ? s.current[_.offset] : 0 }),
+      y = n === `virtual`;
+    return b(`div`, a({ ref: s, css: [Ni, Ii[t], y && Pi, f && Fi], style: v }, Ri));
   };
 function Bi(e, t) {
   var n = Object.keys(e);
@@ -3702,7 +3702,7 @@ function Vi(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Bi(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -3747,7 +3747,7 @@ function Wi(e) {
     i = e.direction,
     a = e.isDropDisabled,
     o = (0, M.useRef)(null),
-    s = N(),
+    s = P(),
     c = s.contextId,
     l = s.getDragState,
     u = l();
@@ -3787,72 +3787,72 @@ function Wi(e) {
 function Gi(e) {
   var t = e.children,
     n = e.droppableId,
-    i = e.type,
-    a = i === void 0 ? `DEFAULT` : i,
-    o = e.direction,
-    s = o === void 0 ? `vertical` : o,
+    r = e.type,
+    i = r === void 0 ? `DEFAULT` : r,
+    a = e.direction,
+    o = a === void 0 ? `vertical` : a,
     c = e.mode,
     l = c === void 0 ? `standard` : c,
     u = e.renderClone,
     d = e.getContainerForClone,
     f = e.isDropDisabled,
-    p = f === void 0 ? !1 : f,
-    m = Jr(p),
-    h = N(),
-    g = h.contextId,
-    _ = h.droppableRegistry,
-    v = Se({ contextId: g, droppableId: n, getIsDropDisabled: m }),
-    b = (0, M.useRef)(null),
-    x = (0, M.useCallback)(
+    m = f === void 0 ? !1 : f,
+    h = Jr(m),
+    _ = P(),
+    v = _.contextId,
+    y = _.droppableRegistry,
+    b = Se({ contextId: v, droppableId: n, getIsDropDisabled: h }),
+    x = (0, M.useRef)(null),
+    S = (0, M.useCallback)(
       function (e) {
         (e &&
           ae(
             e,
-            C(
-              C(C(C({}, L.droppable.type, a), L.droppable.direction, s), I.droppable.id, n),
+            s(
+              s(s(s({}, L.droppable.type, i), L.droppable.direction, o), I.droppable.id, n),
               I.droppable.contextId,
-              g,
+              v,
             ),
           ),
-          (b.current = e));
+          (x.current = e));
       },
-      [g, s, n, a],
+      [v, o, n, i],
     ),
-    S = r((0, M.useReducer)(Ui, Hi), 2),
-    w = S[0],
-    T = S[1],
+    C = g((0, M.useReducer)(Ui, Hi), 2),
+    w = C[0],
+    T = C[1],
     E = w.draggingFromThisWith,
     D = w.draggingOverWith,
     O = w.isDraggingOver,
     k = Ur();
   (0, M.useEffect)(
     function () {
-      var e = b.current;
+      var e = x.current;
       return (
-        y(e instanceof HTMLElement, "innerRef must provide an `HTMLElement`"),
+        p(e instanceof HTMLElement, "innerRef must provide an `HTMLElement`"),
         U(
-          _.register({
+          y.register({
             droppableId: n,
-            type: a,
-            isDropDisabled: p,
+            type: i,
+            isDropDisabled: m,
             parentDroppableId: k,
             element: e,
-            direction: s,
+            direction: o,
             mode: l,
           }),
           Vt({
             element: e,
             getData: function (t) {
               var n = t.input;
-              return ge(v, {
+              return ge(b, {
                 element: e,
                 input: n,
-                allowedEdges: s === `vertical` ? [`top`, `bottom`] : [`left`, `right`],
+                allowedEdges: o === `vertical` ? [`top`, `bottom`] : [`left`, `right`],
               });
             },
             canDrop: function (e) {
               var t = e.source;
-              return !V(t.data) || p ? !1 : t.data.contextId === g && t.data.type === a;
+              return !V(t.data) || m ? !1 : t.data.contextId === v && t.data.type === i;
             },
             onDragLeave: function () {
               T({ type: `DRAG_CLEAR` });
@@ -3861,13 +3861,13 @@ function Gi(e) {
         )
       );
     },
-    [v, n, g, p, a, _, k, s, l],
+    [b, n, v, m, i, y, k, o, l],
   );
   var A = br();
   (0, M.useEffect)(
     function () {
       function e(e) {
-        var t = e.type === a,
+        var t = e.type === i,
           r = e.destination?.droppableId === n;
         return t && ((!O && r) || (O && !r));
       }
@@ -3886,25 +3886,25 @@ function Gi(e) {
         },
       });
     },
-    [n, O, A, a],
+    [n, O, A, i],
   );
   var j = (0, M.useMemo)(
       function () {
-        return O ? M.createElement(zi, { direction: s, mode: l }) : null;
+        return O ? M.createElement(zi, { direction: o, mode: l }) : null;
       },
-      [s, O, l],
+      [o, O, l],
     ),
-    ee = (0, M.useMemo)(
+    N = (0, M.useMemo)(
       function () {
         return {
-          innerRef: x,
-          droppableProps: C(C({}, I.droppable.contextId, g), I.droppable.id, n),
+          innerRef: S,
+          droppableProps: s(s({}, I.droppable.contextId, v), I.droppable.id, n),
           placeholder: l === `standard` ? j : null,
         };
       },
-      [g, j, n, l, x],
+      [v, j, n, l, S],
     ),
-    te = (0, M.useMemo)(
+    ee = (0, M.useMemo)(
       function () {
         return {
           draggingFromThisWith: E,
@@ -3915,67 +3915,67 @@ function Gi(e) {
       },
       [E, D, O],
     ),
-    P = b.current,
-    F = O && l === `virtual` && P;
+    F = x.current,
+    te = O && l === `virtual` && F;
   rr(
     function () {
-      if (F && window.getComputedStyle(P).position === `static`) {
-        var e = P.style.position;
+      if (te && window.getComputedStyle(F).position === `static`) {
+        var e = F.style.position;
         return (
-          (P.style.position = `relative`),
+          (F.style.position = `relative`),
           function () {
-            P.style.position = e;
+            F.style.position = e;
           }
         );
       }
     },
-    [P, F],
+    [F, te],
   );
   var ne = !!u,
     re = (0, M.useMemo)(
       function () {
         return {
-          direction: s,
+          direction: o,
           droppableId: n,
           shouldRenderCloneWhileDragging: ne,
-          isDropDisabled: p,
-          type: a,
+          isDropDisabled: m,
+          type: i,
           mode: l,
         };
       },
-      [s, n, ne, p, a, l],
+      [o, n, ne, m, i, l],
     ),
-    ie = E && l === `virtual` && P;
+    ie = E && l === `virtual` && F;
   return M.createElement(
     Vr,
     { value: re },
-    t(ee, te),
-    F && (0, X.createPortal)(j, P),
+    t(N, ee),
+    te && (0, X.createPortal)(j, F),
     ie &&
       (0, X.createPortal)(
         M.createElement(Wi, {
           droppableId: n,
           draggableId: E,
-          type: a,
-          direction: s,
-          isDropDisabled: p,
+          type: i,
+          direction: o,
+          isDropDisabled: m,
         }),
-        P,
+        F,
       ),
-    u && M.createElement(bi, { droppableId: n, type: a, getContainerForClone: d }, u),
+    u && M.createElement(bi, { droppableId: n, type: i, getContainerForClone: d }, u),
   );
 }
-l();
+i();
 var Ki = [`isRanking`, `isRankingItem`, `testId`],
   qi = (0, M.forwardRef)(function (e, t) {
     var n = e.isRanking,
       r = e.isRankingItem,
       i = e.testId,
-      a = w(e, Ki);
+      s = c(e, Ki);
     return M.createElement(
-      _,
-      s({ ref: t, testId: i }, a, {
-        className: S([
+      y,
+      a({ ref: t, testId: i }, s, {
+        className: o([
           n && `_1e0c1ule`,
           r && `_2rkolb4i _bfhk1bhr _16qs130s`,
           `_12y3e4h9 _mizu1v1w _ra3xnqa1`,
@@ -3983,17 +3983,17 @@ var Ki = [`isRanking`, `isRankingItem`, `testId`],
       }),
     );
   });
-l();
+i();
 var Ji = [`isRanking`, `innerRef`],
   Yi = function (e) {
     var t = e.isRanking,
       n = e.innerRef,
-      r = w(e, Ji);
-    return M.createElement(m, s({}, r, { innerRef: n, className: S([t && `_vchhusvi`]) }));
+      r = c(e, Ji);
+    return M.createElement(h, a({}, r, { innerRef: n, className: o([t && `_vchhusvi`]) }));
   },
   Xi = [`content`, `testId`];
 function Zi(e, t, n) {
-  return ((t = a(t)), i(e, Qi() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
+  return ((t = T(t)), d(e, Qi() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
 function Qi() {
   try {
@@ -4003,14 +4003,14 @@ function Qi() {
     return !!e;
   })();
 }
-var $i = b(
+var $i = _(
     (function (e) {
       function t() {
-        return (g(this, t), Zi(this, t, arguments));
+        return (C(this, t), Zi(this, t, arguments));
       }
       return (
-        c(t, e),
-        o(t, [
+        S(t, e),
+        r(t, [
           {
             key: `render`,
             value: function () {
@@ -4019,31 +4019,31 @@ var $i = b(
                 n = e.head,
                 r = e.isFixedSize,
                 i = e.isRanking,
-                a = e.refWidth,
-                o = e.innerRef,
-                c = e.testId,
-                l = t.content,
-                u = t.testId,
-                d = w(t, Xi),
-                f = n || {},
-                p = f.shouldTruncate,
-                m = f.width,
-                h = v(i, a);
+                o = e.refWidth,
+                s = e.innerRef,
+                l = e.testId,
+                u = t.content,
+                d = t.testId,
+                f = c(t, Xi),
+                p = n || {},
+                m = p.shouldTruncate,
+                h = p.width,
+                g = x(i, o);
               return M.createElement(
                 Yi,
-                s({}, d, {
+                a({}, f, {
                   isFixedSize: r,
-                  shouldTruncate: p,
-                  width: m,
+                  shouldTruncate: m,
+                  width: h,
                   isRanking: i,
-                  style: h,
+                  style: g,
                   onKeyDown: function (e) {
                     return e.stopPropagation();
                   },
-                  innerRef: o,
-                  "data-testid": u || (c && `${c}--rankable--table--body--cell`),
+                  innerRef: s,
+                  "data-testid": d || (l && `${l}--rankable--table--body--cell`),
                 }),
-                l,
+                u,
               );
             },
           },
@@ -4069,7 +4069,7 @@ function na(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ta(Object(n), !0).forEach(function (t) {
-          C(e, t, n[t]);
+          s(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -4080,7 +4080,7 @@ function na(e) {
   return e;
 }
 function ra(e, t, n) {
-  return ((t = a(t)), i(e, ia() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
+  return ((t = T(t)), d(e, ia() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
 function ia() {
   try {
@@ -4090,15 +4090,15 @@ function ia() {
     return !!e;
   })();
 }
-var aa = b(
+var aa = _(
   (function (e) {
     function t() {
       var e;
-      g(this, t);
+      C(this, t);
       var n = [...arguments];
       return (
         (e = ra(this, t, [].concat(n))),
-        C(e, `innerRef`, function (t) {
+        s(e, `innerRef`, function (t) {
           return function (n) {
             (t(n), typeof e.props.innerRef == `function` && e.props.innerRef(n));
           };
@@ -4107,8 +4107,8 @@ var aa = b(
       );
     }
     return (
-      c(t, e),
-      o(t, [
+      S(t, e),
+      r(t, [
         {
           key: `render`,
           value: function () {
@@ -4117,48 +4117,48 @@ var aa = b(
               n = t.row,
               r = t.head,
               i = t.isFixedSize,
-              a = t.isRanking,
-              o = t.refWidth,
-              c = t.rowIndex,
-              l = t.isRankingDisabled,
-              u = t.isHighlighted,
-              d = t.testId,
-              f = n.cells,
-              p = n.testId,
-              m = n.key,
-              h = n.isHighlighted,
-              g = w(n, ea),
-              _ = v(a, o);
-            if (typeof m != `string` && !l)
+              o = t.isRanking,
+              s = t.refWidth,
+              l = t.rowIndex,
+              u = t.isRankingDisabled,
+              d = t.isHighlighted,
+              f = t.testId,
+              p = n.cells,
+              m = n.testId,
+              h = n.key,
+              g = n.isHighlighted,
+              _ = c(n, ea),
+              v = x(o, s);
+            if (typeof h != `string` && !u)
               throw Error(
                 `dynamic-table: ranking is not possible because table row does not have a key. Add the key to the row or disable ranking.`,
               );
             return M.createElement(
               _i,
-              { draggableId: m || c.toString(), index: c, isDragDisabled: l },
+              { draggableId: h || l.toString(), index: l, isDragDisabled: u },
               function (t, n) {
                 return M.createElement(
                   qi,
-                  s({}, g, t.dragHandleProps, t.draggableProps, {
+                  a({}, _, t.dragHandleProps, t.draggableProps, {
                     role: void 0,
                     "aria-labelledby": void 0,
                     "aria-describedby": t.dragHandleProps?.[`aria-describedby`],
                     ref: e.innerRef(t.innerRef),
-                    style: na(na({}, t.draggableProps.style), _),
-                    isHighlighted: u || h,
-                    isRanking: a,
+                    style: na(na({}, t.draggableProps.style), v),
+                    isHighlighted: d || g,
+                    isRanking: o,
                     isRankingItem: n.isDragging,
-                    testId: p || (d && `${d}--rankable--table--body--row`),
+                    testId: m || (f && `${f}--rankable--table--body--row`),
                   }),
-                  f.map(function (e, t) {
+                  p.map(function (e, t) {
                     var n = (r || { cells: [] }).cells[t];
                     return M.createElement($i, {
                       head: n,
                       cell: e,
-                      isRanking: a,
+                      isRanking: o,
                       key: e.key || t,
                       isFixedSize: i,
-                      testId: d,
+                      testId: f,
                     });
                   }),
                 );
@@ -4171,7 +4171,7 @@ var aa = b(
   })(M.Component),
 );
 function oa(e, t, n) {
-  return ((t = a(t)), i(e, sa() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
+  return ((t = T(t)), d(e, sa() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
 function sa() {
   try {
@@ -4195,16 +4195,16 @@ var ca = function (e, t) {
   la = (function (e) {
     function t() {
       var e;
-      g(this, t);
+      C(this, t);
       var n = [...arguments];
       return (
         (e = oa(this, t, [].concat(n))),
-        C(e, `onBeforeDragStart`, function (t) {
+        s(e, `onBeforeDragStart`, function (t) {
           var n = t.draggableId,
             r = { index: t.source.index, key: n };
           e.props.onRankStart(r);
         }),
-        C(e, `onDragEnd`, function (t) {
+        s(e, `onDragEnd`, function (t) {
           var n = e.props,
             r = n.pageRows,
             i = n.onRankEnd,
@@ -4215,8 +4215,8 @@ var ca = function (e, t) {
       );
     }
     return (
-      c(t, e),
-      o(t, [
+      S(t, e),
+      r(t, [
         {
           key: `render`,
           value: function () {
@@ -4225,8 +4225,8 @@ var ca = function (e, t) {
               n = e.pageRows,
               r = e.head,
               i = e.isFixedSize,
-              a = e.isRanking,
-              o = e.isRankingDisabled,
+              o = e.isRanking,
+              s = e.isRankingDisabled,
               c = e.testId,
               l = e.forwardedRef;
             return M.createElement(
@@ -4234,11 +4234,11 @@ var ca = function (e, t) {
               { onBeforeDragStart: this.onBeforeDragStart, onDragEnd: this.onDragEnd },
               M.createElement(
                 Gi,
-                { droppableId: `dynamic-table-droppable`, isDropDisabled: o },
+                { droppableId: `dynamic-table-droppable`, isDropDisabled: s },
                 function (e) {
                   return M.createElement(
                     `tbody`,
-                    s(
+                    a(
                       {
                         "data-testid": c,
                         ref: function (t) {
@@ -4251,12 +4251,12 @@ var ca = function (e, t) {
                     n.map(function (e, n) {
                       return M.createElement(aa, {
                         head: r,
-                        isRanking: a,
+                        isRanking: o,
                         isFixedSize: i,
                         key: e.key,
                         rowIndex: n,
                         row: e,
-                        isRankingDisabled: o,
+                        isRankingDisabled: s,
                         isHighlighted:
                           t !== void 0 && (typeof t == `number` ? t === n : t.indexOf(n) > -1),
                         testId: c && `${c}--${e.key}--rankable--table--row`,
@@ -4272,9 +4272,9 @@ var ca = function (e, t) {
       ])
     );
   })(M.Component),
-  ua = p(
+  ua = u(
     M.forwardRef(function (e, t) {
-      return M.createElement(la, s({}, e, { forwardedRef: t }));
+      return M.createElement(la, a({}, e, { forwardedRef: t }));
     }),
   );
 export { ua as default };
