+++ examples/forge-sql-orm-example-ai/static/forge-orm-example/build-static/assets/custom-theme-D3BYTtIW.js
@@ -10,7 +10,7 @@ import {
   u as c,
   w as l,
   x as u,
-} from "./index-Bk7SZVnx.js";
+} from "./index-BrObFP80.js";
 var d = {
   "color.text.brand": `#669DF1`,
   "elevation.surface.overlay": `#2B2C2F`,
