+++ examples/forge-sql-orm-example-ai/static/forge-orm-example/build-static/assets/index-BrObFP80.js
@@ -5,7 +5,7 @@ const __vite__mapDeps = (
     (m.f = [
       "./iframe-resizer-CDJ7O0mA.js",
       "./rolldown-runtime-BM3Ffeng.js",
-      "./body-BaS5SDy1.js",
+      "./body-Dxgu0KkV.js",
       "./react-dom-vendor-0VJF9GAl.js",
       "./body-BMQTJ_qR.css",
     ]),
@@ -7022,7 +7022,7 @@ var Xi = (function () {
                                       return (
                                         (e.next = 2),
                                         wi(
-                                          () => import(`./custom-theme-B6YGjjYY.js`),
+                                          () => import(`./custom-theme-D3BYTtIW.js`),
                                           [],
                                           import.meta.url,
                                         )
@@ -7760,9 +7760,9 @@ var ja =
             Ne && Pa.disabled,
             Ne && je !== `default` && je !== `subtle` && Pa.sharedDisabled,
             Ne && je === `default` && Fa.disabled,
-            h && !ke && Pa.circle,
             Me === `compact` && Pa.spacingCompact,
             Me === `compact` && yn(`platform-dst-shape-theme-default`) && Pa.spacingCompactT26Shape,
+            h && !ke && Pa.circle,
             S && Pa.fullWidth,
             _ && Pa.buttonIconBefore,
             b && Pa.buttonIconAfter,
@@ -19750,7 +19750,7 @@ var Ry = function (e) {
     );
   },
   zy = (0, d.lazy)(function () {
-    return wi(() => import(`./body-BaS5SDy1.js`), __vite__mapDeps([2, 1, 3, 4]), import.meta.url);
+    return wi(() => import(`./body-Dxgu0KkV.js`), __vite__mapDeps([2, 1, 3, 4]), import.meta.url);
   }),
   By = (0, d.forwardRef)(function (e, t) {
     var n = e.isRankable,
