+++ examples/forge-sql-orm-example-ai/static/forge-orm-example/build-static/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title>forge-sql-orm example</title>
-    <script type="module" crossorigin src="./assets/index-Bk7SZVnx.js"></script>
+    <script type="module" crossorigin src="./assets/index-BrObFP80.js"></script>
     <link rel="modulepreload" crossorigin href="./assets/rolldown-runtime-BM3Ffeng.js" />
     <link rel="modulepreload" crossorigin href="./assets/client-core-vendor-CEzMfQaL.js" />
     <link rel="modulepreload" crossorigin href="./assets/lodash-vendor-Bba9ebez.js" />
