+++ examples/forge-sql-orm-example-ai/static/forge-orm-example/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@atlaskit/button": "^23.11.4",
+    "@atlaskit/button": "^23.11.5",
     "@atlaskit/css-reset": "^7.4.5",
     "@atlaskit/dynamic-table": "^18.4.1",
     "@atlaskit/primitives": "^19.0.1",
