+++ examples/forge-sql-orm-example-backend-ai/static/forge-orm-example/build-static/assets/body-BVfE566-.js
@@ -24,7 +24,7 @@ import {
   w as C,
   x as w,
   y as T,
-} from "./index-CZy_f9dN.js";
+} from "./index-BbaSjtyW.js";
 function E(e, t, n) {
   return ((t = T(t)), u(e, D() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
 }
