+++ examples/forge-sql-orm-example-backend-ai/static/forge-orm-example/build-static/assets/custom-theme-Xz7VWQnJ.js
@@ -10,7 +10,7 @@ import {
   s as c,
   w as l,
   x as u,
-} from "./index-CZy_f9dN.js";
+} from "./index-BbaSjtyW.js";
 var d = {
   "color.text.brand": `#669DF1`,
   "elevation.surface.overlay": `#2B2C2F`,
