+++ examples/forge-sql-orm-example-backend-ai/static/forge-orm-example/build-static/assets/index-BbaSjtyW.js
@@ -5,7 +5,7 @@ const __vite__mapDeps = (
     (m.f = [
       "./iframe-resizer-BMg8MKLf.js",
       "./rolldown-runtime-BYbx6iT9.js",
-      "./body-DY6Q_-9Y.js",
+      "./body-BVfE566-.js",
       "./react-dom-vendor-BHIljUNy.js",
       "./body-BMQTJ_qR.css",
     ]),
@@ -12672,7 +12672,7 @@ var hd = (function () {
                                       return (
                                         (e.next = 2),
                                         Ke(
-                                          () => import(`./custom-theme-CbjR-Q-s.js`),
+                                          () => import(`./custom-theme-Xz7VWQnJ.js`),
                                           [],
                                           import.meta.url,
                                         )
@@ -17938,7 +17938,7 @@ var w_ = function (e) {
     );
   },
   T_ = (0, u.lazy)(function () {
-    return Ke(() => import(`./body-DY6Q_-9Y.js`), __vite__mapDeps([2, 1, 3, 4]), import.meta.url);
+    return Ke(() => import(`./body-BVfE566-.js`), __vite__mapDeps([2, 1, 3, 4]), import.meta.url);
   }),
   E_ = (0, u.forwardRef)(function (e, t) {
     var n = e.isRankable,
@@ -19015,9 +19015,9 @@ var kv =
             De && Mv.disabled,
             De && Te !== `default` && Te !== `subtle` && Mv.sharedDisabled,
             De && Te === `default` && Nv.disabled,
-            h && !Ce && Mv.circle,
             Ee === `compact` && Mv.spacingCompact,
             Ee === `compact` && X(`platform-dst-shape-theme-default`) && Mv.spacingCompactT26Shape,
+            h && !Ce && Mv.circle,
             x && Mv.fullWidth,
             _ && Mv.buttonIconBefore,
             y && Mv.buttonIconAfter,
