+++ examples/forge-sql-orm-example-backend-ai/static/forge-orm-example/build-static/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title>forge-sql-orm example</title>
-    <script type="module" crossorigin src="./assets/index-CZy_f9dN.js"></script>
+    <script type="module" crossorigin src="./assets/index-BbaSjtyW.js"></script>
     <link rel="modulepreload" crossorigin href="./assets/rolldown-runtime-BYbx6iT9.js" />
     <link rel="modulepreload" crossorigin href="./assets/client-core-vendor-C5fkS-7z.js" />
     <link rel="modulepreload" crossorigin href="./assets/lodash-vendor-CY-a7-Db.js" />
