+++ examples/forge-sql-orm-example-optimistic-locking/static/forge-orm-example/build-static/assets/body-D9h5kDZ1.js
@@ -1,5 +1,5 @@
-import { o as e } from "./rolldown-runtime-BM3Ffeng.js";
-import { n as t, r as n } from "./react-dom-vendor-0VJF9GAl.js";
+import { a as e } from "./rolldown-runtime-BYbx6iT9.js";
+import { n as t, r as n } from "./react-dom-vendor-BHIljUNy.js";
 import {
   C as r,
   D as i,
@@ -10,23 +10,23 @@ import {
   _ as l,
   a as u,
   b as d,
-  c as f,
+  f,
   g as p,
   h as m,
   i as h,
   k as g,
-  n as _,
-  o as v,
-  r as y,
-  s as b,
+  m as _,
+  n as v,
+  p as y,
+  r as b,
   t as x,
   v as S,
   w as C,
   x as w,
   y as T,
-} from "./index-Bk7SZVnx.js";
+} from "./index-RNH_WMbf.js";
 function E(e, t, n) {
-  return ((t = T(t)), d(e, D() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
+  return ((t = a(t)), i(e, D() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
 }
 function D() {
   try {
@@ -41,11 +41,11 @@ var O = !0,
   A = (function (e) {
     function t(e) {
       var n;
-      return (C(this, t), (n = E(this, t)), (n.message = e), n);
+      return (g(this, t), (n = E(this, t)), (n.message = e), n);
     }
     return (
-      S(t, e),
-      r(t, [
+      c(t, e),
+      o(t, [
         {
           key: `toString`,
           value: function () {
@@ -54,17 +54,17 @@ var O = !0,
         },
       ])
     );
-  })(l(Error));
+  })(T(Error));
 function j(e, t) {
   if (!e) throw O ? new A(k) : new A(`${k}: ${t || ``}`);
 }
 var M = e(n()),
-  N = (0, M.createContext)(null);
-function P() {
-  var e = (0, M.useContext)(N);
+  ee = (0, M.createContext)(null);
+function N() {
+  var e = (0, M.useContext)(ee);
   return (j(e !== null, `Unable to find DragDropContext context`), e);
 }
-function ee(e) {
+function te(e) {
   var t = e.children,
     n = e.contextId,
     r = e.getDragState,
@@ -76,13 +76,13 @@ function ee(e) {
       },
       [n, r, i, a],
     );
-  return M.createElement(N.Provider, { value: o }, t);
+  return M.createElement(ee.Provider, { value: o }, t);
 }
-function F(e) {
+function P(e) {
   return { margin: window.getComputedStyle(e).margin, rect: e.getBoundingClientRect() };
 }
-function te() {
-  var e = P().getDragState,
+function F() {
+  var e = N().getDragState,
     t = e();
   return t.isDragging ? t.draggableDimensions : null;
 }
@@ -99,7 +99,7 @@ function re() {
   };
 }
 function ie() {
-  var e = g((0, M.useState)(re), 1)[0];
+  var e = r((0, M.useState)(re), 1)[0];
   return (
     (0, M.useEffect)(
       function () {
@@ -132,11 +132,11 @@ function R(e, t) {
   return (j(n !== null, `Expected '${t}' to be present`), n);
 }
 function ae(e, t) {
-  for (var n = 0, r = Object.entries(t); n < r.length; n++) {
-    var i = g(r[n], 2),
-      a = i[0],
-      o = i[1];
-    e.setAttribute(a, o);
+  for (var n = 0, i = Object.entries(t); n < i.length; n++) {
+    var a = r(i[n], 2),
+      o = a[0],
+      s = a[1];
+    e.setAttribute(o, s);
   }
   return function () {
     for (var n = 0, r = Object.keys(t); n < r.length; n++) {
@@ -203,7 +203,7 @@ function pe(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? fe(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -234,7 +234,7 @@ function ge(e, t) {
     i = t.allowedEdges,
     a = { x: r.clientX, y: r.clientY },
     o = n.getBoundingClientRect(),
-    c =
+    s =
       i
         .map(function (e) {
           return { edge: e, value: me[e](o, a) };
@@ -242,7 +242,7 @@ function ge(e, t) {
         .sort(function (e, t) {
           return e.value - t.value;
         })[0]?.edge ?? null;
-  return pe(pe({}, e), {}, s({}, he, c));
+  return pe(pe({}, e), {}, C({}, he, s));
 }
 function _e(e) {
   return e[he] ?? null;
@@ -259,9 +259,9 @@ function ye(e) {
     a = e.type;
   return (0, M.useMemo)(
     function () {
-      return s(
-        s(
-          s(s(s(s({}, ve, !0), `draggableId`, t), `droppableId`, n), `getIndex`, r),
+      return C(
+        C(
+          C(C(C(C({}, ve, !0), `draggableId`, t), `droppableId`, n), `getIndex`, r),
           `contextId`,
           i,
         ),
@@ -282,7 +282,7 @@ function Se(e) {
     r = e.getIsDropDisabled;
   return (0, M.useMemo)(
     function () {
-      return s(s(s(s({}, be, !0), `contextId`, t), `droppableId`, n), `getIsDropDisabled`, r);
+      return C(C(C(C({}, be, !0), `contextId`, t), `droppableId`, n), `getIsDropDisabled`, r);
     },
     [t, n, r],
   );
@@ -300,7 +300,7 @@ var we = [`droppableId`, `getIndex`],
 function Ee(e) {
   var t = e.droppableId,
     n = e.getIndex,
-    r = c(e, we),
+    r = w(e, we),
     i = n(),
     a = _e(r);
   return ((a === `bottom` || a === `right`) && (i += 1), { droppableId: t, index: i });
@@ -308,7 +308,7 @@ function Ee(e) {
 function De(e) {
   var t = e.contextId,
     n = e.droppableId,
-    r = c(e, Te),
+    r = w(e, Te),
     i = Ce({ contextId: t, droppableId: n });
   if (i.length === 0) return { droppableId: n, index: 0 };
   var a = _e(r);
@@ -357,7 +357,7 @@ function ke() {
   };
 }
 function Ae() {
-  return g((0, M.useState)(ke), 1)[0];
+  return r((0, M.useState)(ke), 1)[0];
 }
 function U() {
   var e = [...arguments];
@@ -372,10 +372,10 @@ function Me(e) {
   return e instanceof Element && e.hasAttribute(`data-pdnd-honey-pot`);
 }
 function Ne(e) {
-  var t = g(document.elementsFromPoint(e.x, e.y), 2),
+  var t = r(document.elementsFromPoint(e.x, e.y), 2),
     n = t[0],
-    r = t[1];
-  return n ? (Me(n) ? (r ?? null) : n) : null;
+    i = t[1];
+  return n ? (Me(n) ? (i ?? null) : n) : null;
 }
 var Pe = 2147483647,
   Fe = {
@@ -400,7 +400,7 @@ function W(e) {
 var Ie = W(function () {
     return typeof HTMLElement < `u` && typeof HTMLElement.prototype.showPopover == `function`;
   }),
-  G = m();
+  G = f();
 function Le(e, t) {
   var n = Object.keys(e);
   if (Object.getOwnPropertySymbols) {
@@ -418,7 +418,7 @@ function Re(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Le(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -648,7 +648,7 @@ function it(e) {
     pageY: e.pageY,
   };
 }
-var at = f(function (e) {
+var at = x(function (e) {
     return e();
   }),
   ot = (function () {
@@ -760,12 +760,12 @@ function ft(e) {
   function u() {
     (o.current.dropTargets.length && c({ dropTargets: [], input: o.current.input }),
       s.drop({ current: o.current, updatedSourcePayload: null }),
-      d());
+      f());
   }
-  function d() {
-    ((ct.isActive = !1), f());
+  function f() {
+    ((ct.isActive = !1), p());
   }
-  var f = (0, G.bindAll)(
+  var p = (0, G.bindAll)(
     window,
     [
       {
@@ -798,7 +798,7 @@ function ft(e) {
               current: o.current,
               updatedSourcePayload: n.type === `external` ? n.getDropPayload(e) : null,
             }),
-            d());
+            f());
         },
       },
       {
@@ -807,7 +807,7 @@ function ft(e) {
           ((o.current = { dropTargets: o.current.dropTargets, input: it(e) }), u());
         },
       },
-    ].concat(w(rt({ onDragEnd: u }))),
+    ].concat(d(rt({ onDragEnd: u }))),
     { capture: !0 },
   );
   s.start({ nativeSetDragImage: ut(t) });
@@ -869,7 +869,7 @@ function K(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? bt(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -973,19 +973,19 @@ function Tt(e) {
       return i instanceof Node ? c({ source: t, target: i.parentElement, input: o, result: l }) : l;
     var u = i.closest(a);
     if (u == null) return l;
-    var d = r.get(u);
-    if (d == null) return l;
-    var f = { input: o, source: t, element: d.element };
-    if (d.canDrop && !d.canDrop(f))
-      return c({ source: t, target: d.element.parentElement, input: o, result: l });
-    var p = d.getData?.call(d, f) ?? {},
-      m = d.getDropEffect?.call(d, f) ?? n,
-      h = { data: p, element: d.element, dropEffect: m, isActiveDueToStickiness: !1 };
+    var f = r.get(u);
+    if (f == null) return l;
+    var p = { input: o, source: t, element: f.element };
+    if (f.canDrop && !f.canDrop(p))
+      return c({ source: t, target: f.element.parentElement, input: o, result: l });
+    var m = f.getData?.call(f, p) ?? {},
+      h = f.getDropEffect?.call(f, p) ?? n,
+      g = { data: m, element: f.element, dropEffect: h, isActiveDueToStickiness: !1 };
     return c({
       source: t,
-      target: d.element.parentElement,
+      target: f.element.parentElement,
       input: o,
-      result: [].concat(w(l), [h]),
+      result: [].concat(d(l), [g]),
     });
   }
   function l(e) {
@@ -1061,10 +1061,10 @@ function Tt(e) {
       }
     },
   };
-  function d(e) {
+  function f(e) {
     u[e.eventName](e);
   }
-  function f(e) {
+  function p(e) {
     var t = e.source,
       n = e.target,
       i = e.input,
@@ -1090,7 +1090,7 @@ function Tt(e) {
     }
     return wt(u);
   }
-  return { dropTargetForConsumers: s, getIsOver: f, dispatchEvent: d };
+  return { dropTargetForConsumers: s, getIsOver: p, dispatchEvent: f };
 }
 function Et(e, t) {
   var n = (typeof Symbol < `u` && e[Symbol.iterator]) || e[`@@iterator`];
@@ -1172,7 +1172,7 @@ function At(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? kt(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1279,44 +1279,44 @@ var zt = Ye(),
             if (e.canStart(t) && !t.defaultPrevented && t.dataTransfer) {
               var n = t.target;
               if (n instanceof HTMLElement) {
-                var r = Lt.get(n);
-                if (r) {
-                  var i = it(t),
-                    a = { element: r.element, dragHandle: r.dragHandle ?? null, input: i };
-                  if (r.canDrag && !r.canDrag(a)) {
+                var i = Lt.get(n);
+                if (i) {
+                  var a = it(t),
+                    o = { element: i.element, dragHandle: i.dragHandle ?? null, input: a };
+                  if (i.canDrag && !i.canDrag(o)) {
                     t.preventDefault();
                     return;
                   }
-                  if (r.dragHandle) {
-                    var o = Ne({ x: i.clientX, y: i.clientY });
-                    if (!r.dragHandle.contains(o)) {
+                  if (i.dragHandle) {
+                    var s = Ne({ x: a.clientX, y: a.clientY });
+                    if (!i.dragHandle.contains(s)) {
                       t.preventDefault();
                       return;
                     }
                   }
-                  var s = r.getInitialDataForExternal?.call(r, a) ?? null;
-                  if (s)
-                    for (var c = 0, l = Object.entries(s); c < l.length; c++) {
-                      var u = g(l[c], 2),
-                        d = u[0],
-                        f = u[1];
-                      t.dataTransfer.setData(d, f ?? ``);
+                  var c = i.getInitialDataForExternal?.call(i, o) ?? null;
+                  if (c)
+                    for (var l = 0, u = Object.entries(c); l < u.length; l++) {
+                      var d = r(u[l], 2),
+                        f = d[0],
+                        p = d[1];
+                      t.dataTransfer.setData(f, p ?? ``);
                     }
                   (Nt() &&
                     !t.dataTransfer.types.includes(`text/plain`) &&
                     !t.dataTransfer.types.includes(`text/uri-list`) &&
                     t.dataTransfer.setData(Ft, Pt),
                     t.dataTransfer.setData(It, ``));
-                  var p = {
+                  var m = {
                     type: `element`,
                     payload: {
-                      element: r.element,
-                      dragHandle: r.dragHandle ?? null,
-                      data: r.getInitialData?.call(r, a) ?? {},
+                      element: i.element,
+                      dragHandle: i.dragHandle ?? null,
+                      data: i.getInitialData?.call(i, o) ?? {},
                     },
                     startedFrom: `internal`,
                   };
-                  e.start({ event: t, dragType: p });
+                  e.start({ event: t, dragType: m });
                 }
               }
             }
@@ -1339,7 +1339,7 @@ function Ut(e) {
   return W(U(Bt.registerUsage(), Rt(e), yt(e.element, { attribute: `draggable`, value: `true` })));
 }
 function Wt(e, t, n) {
-  return ((t = T(t)), d(e, Gt() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
+  return ((t = a(t)), i(e, Gt() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
 }
 function Gt() {
   try {
@@ -1385,11 +1385,11 @@ function Kt(e) {
 }
 var qt = (function (e) {
   function t() {
-    return (C(this, t), Wt(this, t, arguments));
+    return (g(this, t), Wt(this, t, arguments));
   }
   return (
-    S(t, e),
-    r(
+    c(t, e),
+    o(
       t,
       [
         {
@@ -1430,7 +1430,7 @@ function Yt(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Jt(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1534,7 +1534,7 @@ function q(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? rn(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -1679,12 +1679,12 @@ function ln(e) {
           mode: `SNAP`,
         });
         var u = c(),
-          d = R(B({ attribute: I.droppable.id, value: u.droppableId }), L.droppable.direction);
-        j(d === `vertical` || d === `horizontal`);
-        function f() {
+          f = R(B({ attribute: I.droppable.id, value: u.droppableId }), L.droppable.direction);
+        j(f === `vertical` || f === `horizontal`);
+        function p() {
           t.stopDrag({ reason: `CANCEL` });
         }
-        var p = [
+        var m = [
           `mousedown`,
           `mouseup`,
           `click`,
@@ -1693,7 +1693,7 @@ function ln(e) {
           `wheel`,
           `visibilitychange`,
         ].map(function (e) {
-          return { type: e, listener: f };
+          return { type: e, listener: p };
         });
         i(
           (0, G.bindAll)(
@@ -1712,12 +1712,12 @@ function ln(e) {
                       (e.preventDefault(), t.stopDrag({ reason: `CANCEL` }));
                       return;
                     }
-                    (i = (o = cn[d])[e.key]) == null ||
+                    (i = (o = cn[f])[e.key]) == null ||
                       i.call(o, e, { dragController: t, droppableRegistry: n, contextId: r });
                   }
                 },
               },
-            ].concat(w(p)),
+            ].concat(d(m)),
           ),
         );
       },
@@ -2087,7 +2087,7 @@ function Qn() {
           },
         },
         { type: `dragend`, listener: $n },
-      ].concat(w(rt({ onDragEnd: $n }))),
+      ].concat(d(rt({ onDragEnd: $n }))),
       { capture: !1 },
     )));
 }
@@ -2167,7 +2167,7 @@ function ar(e) {
     n = e.styles;
   return `${t} { ${Object.entries(n)
     .map(function (e) {
-      var t = g(e, 2);
+      var t = r(e, 2);
       return `${t[0]}: ${t[1]};`;
     })
     .join(` `)} }`;
@@ -2320,26 +2320,26 @@ function gr() {
   };
 }
 function _r() {
-  return g((0, M.useState)(gr), 1)[0];
+  return r((0, M.useState)(gr), 1)[0];
 }
 var vr = (0, M.createContext)(null);
 function yr(e) {
   var t = e.children,
     n = e.lifecycle,
-    r = (0, M.useCallback)(
+    i = (0, M.useCallback)(
       function (e) {
-        for (var t = [], r = 0, i = Object.entries(e); r < i.length; r++) {
-          var a = i[r],
-            o = g(a, 2),
-            s = o[0],
-            c = o[1];
-          t.push(n.addResponder(s, c));
+        for (var t = [], i = 0, a = Object.entries(e); i < a.length; i++) {
+          var o = a[i],
+            s = r(o, 2),
+            c = s[0],
+            l = s[1];
+          t.push(n.addResponder(c, l));
         }
         return U.apply(void 0, t);
       },
       [n],
     );
-  return M.createElement(vr.Provider, { value: r }, t);
+  return M.createElement(vr.Provider, { value: i }, t);
 }
 function br() {
   var e = (0, M.useContext)(vr);
@@ -2445,7 +2445,7 @@ function Mr() {
   return { schedule: t, flush: n };
 }
 function Nr() {
-  return g((0, M.useState)(Mr), 1)[0];
+  return r((0, M.useState)(Mr), 1)[0];
 }
 var Pr = 0;
 function Fr() {
@@ -2458,22 +2458,22 @@ function Ir(e) {
 function Lr(e) {
   var t = e.children,
     n = e.dragHandleUsageInstructions,
-    r = n === void 0 ? Ar : n,
-    i = e.nonce,
-    a = e.onBeforeCapture,
-    o = e.onBeforeDragStart,
-    s = e.onDragStart,
-    c = e.onDragUpdate,
-    l = e.onDragEnd,
-    u = g((0, M.useState)(Fr), 1)[0];
-  Qt({ contextId: u, text: r });
-  var d = _r(),
-    f = Nr(),
-    p = f.schedule,
-    m = f.flush,
-    h = (0, M.useRef)({ isDragging: !1 }),
+    i = n === void 0 ? Ar : n,
+    a = e.nonce,
+    o = e.onBeforeCapture,
+    s = e.onBeforeDragStart,
+    c = e.onDragStart,
+    l = e.onDragUpdate,
+    u = e.onDragEnd,
+    d = r((0, M.useState)(Fr), 1)[0];
+  Qt({ contextId: d, text: i });
+  var f = _r(),
+    p = Nr(),
+    m = p.schedule,
+    h = p.flush,
+    g = (0, M.useRef)({ isDragging: !1 }),
     _ = (0, M.useCallback)(function () {
-      return h.current;
+      return g.current;
     }, []),
     v = Ae(),
     y = (0, M.useCallback)(
@@ -2502,35 +2502,35 @@ function Lr(e) {
         var t = e.targetLocation,
           n = e.isImmediate,
           r = n === void 0 ? !1 : n;
-        if (!h.current.isDragging) return;
-        var i = h.current,
+        if (!g.current.isDragging) return;
+        var i = g.current,
           a = i.prevDestination,
           o = i.draggableId,
           s = i.type,
-          l = i.sourceLocation,
-          u = Xt({ start: l, target: t });
+          c = i.sourceLocation,
+          u = Xt({ start: c, target: t });
         if (H(a, u)) return;
-        Object.assign(h.current, { prevDestination: u, sourceLocation: l, targetLocation: t });
-        var f = {
-            mode: h.current.mode,
+        Object.assign(g.current, { prevDestination: u, sourceLocation: c, targetLocation: t });
+        var d = {
+            mode: g.current.mode,
             draggableId: o,
             type: s,
-            source: l,
+            source: c,
             destination: u,
             combine: null,
           },
-          m = t ? v.getEntry({ droppableId: t.droppableId }) : null;
-        (d.dispatch(`onPrePendingDragUpdate`, { update: f, targetLocation: t }),
-          d.dispatch(`onPendingDragUpdate`, { update: f, targetLocation: t, droppable: m }));
-        function g() {
-          var e = kr(`onDragUpdate`, f),
+          p = t ? v.getEntry({ droppableId: t.droppableId }) : null;
+        (f.dispatch(`onPrePendingDragUpdate`, { update: d, targetLocation: t }),
+          f.dispatch(`onPendingDragUpdate`, { update: d, targetLocation: t, droppable: p }));
+        function h() {
+          var e = kr(`onDragUpdate`, d),
             t = e.provided,
             n = e.getMessage;
-          (c?.(f, t), Tr(n()));
+          (l?.(d, t), Tr(n()));
         }
-        r ? g() : p(g);
+        r ? h() : m(h);
       },
-      [v, d, c, p],
+      [v, f, l, m],
     ),
     x = (0, M.useCallback)(
       function (e) {
@@ -2538,40 +2538,40 @@ function Lr(e) {
           n = e.type,
           r = e.getSourceLocation,
           i = e.sourceElement,
-          c = e.mode;
-        if (!h.current.isDragging) {
-          a?.({ draggableId: t, mode: c });
-          var l = { mode: c, draggableId: t, type: n, source: r() },
+          a = e.mode;
+        if (!g.current.isDragging) {
+          o?.({ draggableId: t, mode: a });
+          var l = { mode: a, draggableId: t, type: n, source: r() },
             u = document.activeElement,
-            f =
+            d =
               u instanceof HTMLElement && u.hasAttribute(I.dragHandle.draggableId)
                 ? R(u, I.dragHandle.draggableId)
                 : null,
-            m = l.source.droppableId,
-            g = v.getEntry({ droppableId: m });
-          (j(g, `should have entry for droppable '${m}'`),
-            (h.current = {
+            p = l.source.droppableId,
+            h = v.getEntry({ droppableId: p });
+          (j(h, `should have entry for droppable '${p}'`),
+            (g.current = {
               isDragging: !0,
-              mode: c,
-              draggableDimensions: F(i),
-              restoreFocusTo: f,
+              mode: a,
+              draggableDimensions: P(i),
+              restoreFocusTo: d,
               draggableId: t,
               type: n,
               prevDestination: l.source,
               sourceLocation: l.source,
               targetLocation: l.source,
-              draggableInitialOffsetInSourceDroppable: Ir({ element: i, mode: g.mode }),
+              draggableInitialOffsetInSourceDroppable: Ir({ element: i, mode: h.mode }),
             }),
-            o?.(l),
-            d.dispatch(`onPendingDragStart`, { start: l, droppable: g }),
-            p(function () {
-              var e = { mode: c, draggableId: t, type: n, source: r() },
+            s?.(l),
+            f.dispatch(`onPendingDragStart`, { start: l, droppable: h }),
+            m(function () {
+              var e = { mode: a, draggableId: t, type: n, source: r() },
                 i = kr(`onDragStart`, e),
-                a = i.provided,
-                o = i.getMessage;
-              (s?.(e, a),
-                Tr(o()),
-                p(function () {
+                o = i.provided,
+                s = i.getMessage;
+              (c?.(e, o),
+                Tr(s()),
+                m(function () {
                   var t = e.source.droppableId,
                     n = v.getEntry({ droppableId: t });
                   n != null &&
@@ -2581,46 +2581,46 @@ function Lr(e) {
             }));
         }
       },
-      [v, y, d, a, o, s, p, b],
+      [v, y, f, o, s, c, m, b],
     ),
     S = ie(),
     C = (0, M.useCallback)(
       function (e) {
         var t = e.reason;
-        if (h.current.isDragging) {
+        if (g.current.isDragging) {
           (S.runCleanupFn(), t === `CANCEL` && b({ targetLocation: null }));
-          var n = h.current,
+          var n = g.current,
             r = n.mode,
             i = n.restoreFocusTo,
             a = n.sourceLocation,
             o = n.targetLocation,
             s = n.type,
             c = n.draggableId;
-          ((h.current = { isDragging: !1 }), m());
-          var f = Xt({ start: a, target: o }),
+          ((g.current = { isDragging: !1 }), h());
+          var l = Xt({ start: a, target: o }),
             p = {
-              reason: f === null ? `CANCEL` : `DROP`,
+              reason: l === null ? `CANCEL` : `DROP`,
               type: s,
               source: a,
-              destination: f,
+              destination: l,
               mode: r,
               draggableId: c,
               combine: null,
             };
-          d.dispatch(`onBeforeDragEnd`, { draggableId: c });
-          var g = kr(`onDragEnd`, p),
-            _ = g.provided,
-            v = g.getMessage;
-          (l(p, _),
+          f.dispatch(`onBeforeDragEnd`, { draggableId: c });
+          var m = kr(`onDragEnd`, p),
+            _ = m.provided,
+            v = m.getMessage;
+          (u(p, _),
             Tr(v()),
             i &&
               requestAnimationFrame(function () {
-                var e = ce({ contextId: u, draggableId: c });
+                var e = ce({ contextId: d, draggableId: c });
                 e && e.focus();
               }));
         }
       },
-      [u, m, S, d, l, b],
+      [d, h, S, f, u, b],
     ),
     w = (0, M.useMemo)(
       function () {
@@ -2628,16 +2628,16 @@ function Lr(e) {
       },
       [_, x, C, b],
     );
-  nr({ dragController: w, contextId: u });
+  nr({ dragController: w, contextId: d });
   var T = ln({
       dragController: w,
       droppableRegistry: v,
-      contextId: u,
+      contextId: d,
       setKeyboardCleanupFn: S.setCleanupFn,
     }).startKeyboardDrag,
     E = (0, M.useCallback)(
       function (e) {
-        var t = h.current;
+        var t = g.current;
         t.isDragging &&
           e.isDropDisabled &&
           e.droppableId === t.targetLocation?.droppableId &&
@@ -2647,16 +2647,16 @@ function Lr(e) {
     );
   return (
     v.setUpdateListener(E),
-    lr({ contextId: u, nonce: i }),
+    lr({ contextId: d, nonce: a }),
     M.createElement(
       qt,
-      { contextId: u, dragController: w },
+      { contextId: d, dragController: w },
       M.createElement(
         yr,
-        { lifecycle: d },
+        { lifecycle: f },
         M.createElement(
-          ee,
-          { contextId: u, getDragState: _, startKeyboardDrag: T, droppableRegistry: v },
+          te,
+          { contextId: d, getDragState: _, startKeyboardDrag: T, droppableRegistry: v },
           t,
         ),
       ),
@@ -2721,7 +2721,7 @@ function Wr(e) {
   );
 }
 function Gr() {
-  return { startKeyboardDrag: P().startKeyboardDrag };
+  return { startKeyboardDrag: N().startKeyboardDrag };
 }
 function Kr() {
   return z({ attribute: L.dropIndicator });
@@ -2761,7 +2761,7 @@ function Zr(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Xr(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2839,7 +2839,7 @@ function oi(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ai(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2852,10 +2852,10 @@ function oi(e) {
 var si = (0, M.memo)(
     (0, M.forwardRef)(function (e, t) {
       var n = e.style,
-        r = te(),
-        i = P().contextId,
-        o = s({}, I.placeholder.contextId, i),
-        c = (0, M.useMemo)(
+        r = F(),
+        i = N().contextId,
+        a = C({}, I.placeholder.contextId, i),
+        o = (0, M.useMemo)(
           function () {
             if (r) {
               var e = r.margin,
@@ -2868,7 +2868,7 @@ var si = (0, M.memo)(
           },
           [r, n],
         );
-      return M.createElement(`div`, a({ ref: t, style: c }, o));
+      return M.createElement(`div`, s({ ref: t, style: o }, a));
     }),
   ),
   Q = {
@@ -2924,7 +2924,7 @@ function $(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ci(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -2943,7 +2943,7 @@ var di = {
     var t = Q[e.direction],
       n = t.mainAxis,
       r = t.crossAxis;
-    return s(s({}, n.name, 0), r.name, 24);
+    return C(C({}, n.name, 0), r.name, 24);
   },
   home: function (e) {
     var t = e.droppable.direction,
@@ -2953,8 +2953,8 @@ var di = {
     var i = { x: n.x - r.rect.x, y: n.y - r.rect.y },
       a = Q[t],
       o = a.mainAxis,
-      c = a.crossAxis;
-    return s(s({}, o.name, i[o.name]), c.name, i[c.name] + 24);
+      s = a.crossAxis;
+    return C(C({}, o.name, i[o.name]), s.name, i[s.name] + 24);
   },
   away: function (e) {
     var t = e.droppable.direction,
@@ -2964,8 +2964,8 @@ var di = {
     var i = { x: n.x - r.rect.x, y: n.y - r.rect.y },
       a = Q[t],
       o = a.mainAxis,
-      c = a.crossAxis;
-    return s(s({}, o.name, i[o.name] - 0.5 * r.rect[o.style.length]), c.name, i[c.name] + 24);
+      s = a.crossAxis;
+    return C(C({}, o.name, i[o.name] - 0.5 * r.rect[o.style.length]), s.name, i[s.name] + 24);
   },
 };
 function fi(e, t) {
@@ -3067,53 +3067,53 @@ var gi = function () {};
 function _i(e) {
   var t = e.children,
     n = e.draggableId,
-    r = e.index,
-    i = e.isDragDisabled,
-    a = i === void 0 ? !1 : i,
-    o = e.disableInteractiveElementBlocking,
-    c = o === void 0 ? !1 : o,
+    i = e.index,
+    a = e.isDragDisabled,
+    o = a === void 0 ? !1 : a,
+    s = e.disableInteractiveElementBlocking,
+    c = s === void 0 ? !1 : s,
     l = Hr(),
     u = l.direction,
     d = l.droppableId,
     f = l.type,
-    m = l.mode,
-    h = P(),
-    _ = h.contextId,
-    v = h.getDragState,
-    y = (0, M.useRef)(null),
-    b = (0, M.useRef)(null),
-    x = ie(),
-    S = x.setCleanupFn,
-    C = x.runCleanupFn,
+    p = l.mode,
+    m = N(),
+    h = m.contextId,
+    g = m.getDragState,
+    _ = (0, M.useRef)(null),
+    v = (0, M.useRef)(null),
+    b = ie(),
+    x = b.setCleanupFn,
+    S = b.runCleanupFn,
     w = (0, M.useCallback)(
       function (e) {
-        (y.current && C(),
-          e && S(ae(e, s(s({}, L.draggable.droppableId, d), L.draggable.index, String(r)))),
-          (y.current = e),
-          (b.current = ce({ contextId: _, draggableId: n })));
+        (_.current && S(),
+          e && x(ae(e, C(C({}, L.draggable.droppableId, d), L.draggable.index, String(i)))),
+          (_.current = e),
+          (v.current = ce({ contextId: h, draggableId: n })));
       },
-      [_, n, d, r, C, S],
+      [h, n, d, i, S, x],
     ),
-    T = Jr(r),
-    E = g((0, M.useReducer)(mi, li), 2),
+    T = Jr(i),
+    E = r((0, M.useReducer)(mi, li), 2),
     D = E[0],
     O = E[1],
-    k = ye({ draggableId: n, droppableId: d, getIndex: T, contextId: _, type: f }),
+    k = ye({ draggableId: n, droppableId: d, getIndex: T, contextId: h, type: f }),
     A = D.type === `dragging`,
-    N = D.type === `hiding`,
-    ee = Hr(),
-    F = ee.shouldRenderCloneWhileDragging,
-    ne = ee.isDropDisabled,
+    ee = D.type === `hiding`,
+    te = Hr(),
+    P = te.shouldRenderCloneWhileDragging,
+    ne = te.isDropDisabled,
     re = br(),
     R = Gr().startKeyboardDrag;
   ((0, M.useEffect)(
     function () {
-      if (D.type === `idle` && !a) {
-        var e = y.current;
-        p(e instanceof HTMLElement);
-        var t = b.current;
+      if (D.type === `idle` && !o) {
+        var e = _.current;
+        y(e instanceof HTMLElement);
+        var t = v.current;
         return (
-          p(t instanceof HTMLElement),
+          y(t instanceof HTMLElement),
           (0, G.bindAll)(t, [
             {
               type: `keydown`,
@@ -3137,14 +3137,14 @@ function _i(e) {
         );
       }
     },
-    [c, n, d, T, a, R, D.type, f],
+    [c, n, d, T, o, R, D.type, f],
   ),
     (0, M.useEffect)(
       function () {
-        if (!N && !a) {
-          var e = y.current;
+        if (!ee && !o) {
+          var e = _.current;
           j(e instanceof HTMLElement);
-          var t = b.current;
+          var t = v.current;
           return (
             j(t instanceof HTMLElement),
             Ut({
@@ -3166,11 +3166,11 @@ function _i(e) {
           );
         }
       },
-      [k, c, a, A, N],
+      [k, c, o, A, ee],
     ));
-  var oe = D.type !== `idle` && m === `standard`,
+  var oe = D.type !== `idle` && p === `standard`,
     z = (0, M.useRef)(null);
-  Wr({ elementRef: oe ? z : y, data: k, direction: u, contextId: _, isDropDisabled: ne, type: f });
+  Wr({ elementRef: oe ? z : _, data: k, direction: u, contextId: h, isDropDisabled: ne, type: f });
   var se = (0, M.useRef)(!0);
   ((0, M.useEffect)(function () {
     return (
@@ -3182,18 +3182,18 @@ function _i(e) {
   }, []),
     (0, M.useEffect)(
       function () {
-        var e = v();
-        F &&
+        var e = g();
+        P &&
           e.isDragging &&
           e.draggableId === k.draggableId &&
           O({ type: `START_HIDING`, payload: { mode: e.mode } });
       },
-      [k.draggableId, v, F],
+      [k.draggableId, g, P],
     ));
-  var B = te();
+  var B = F();
   (0, M.useEffect)(
     function () {
-      return F
+      return P
         ? re({
             onPendingDragStart: function (e) {
               var t = e.start;
@@ -3213,7 +3213,7 @@ function _i(e) {
                   if (t.mode === `FLUID`)
                     return O({ type: `START_POINTER_DRAG`, payload: { start: t } });
                   if (t.mode === `SNAP`) {
-                    var r = v();
+                    var r = g();
                     return (
                       j(r.isDragging && r.draggableDimensions),
                       O({
@@ -3235,8 +3235,8 @@ function _i(e) {
                   (O({ type: `UPDATE_DRAG`, payload: { update: t } }),
                   t.mode === `SNAP` &&
                     queueMicrotask(function () {
-                      if (v().isDragging) {
-                        var e = qr(_),
+                      if (g().isDragging) {
+                        var e = qr(h),
                           r = e ? e.getBoundingClientRect() : null,
                           i = Kr();
                         O({
@@ -3271,21 +3271,21 @@ function _i(e) {
             }),
           );
     },
-    [k.draggableId, k.contextId, re, F, u, _, B, v],
+    [k.draggableId, k.contextId, re, P, u, h, B, g],
   );
   var le = (0, M.useMemo)(
       function () {
         return {
-          draggableProps: s(
-            s(s({}, I.draggable.contextId, _), I.draggable.id, n),
+          draggableProps: C(
+            C(C({}, I.draggable.contextId, h), I.draggable.id, n),
             `style`,
             ti({ draggableDimensions: B, draggableState: D }),
           ),
-          dragHandleProps: s(
-            s(
-              s(
-                s(
-                  s({ role: `button`, "aria-describedby": Zt(_) }, I.dragHandle.contextId, _),
+          dragHandleProps: C(
+            C(
+              C(
+                C(
+                  C({ role: `button`, "aria-describedby": Zt(h) }, I.dragHandle.contextId, h),
                   I.dragHandle.draggableId,
                   n,
                 ),
@@ -3301,19 +3301,19 @@ function _i(e) {
           innerRef: w,
         };
       },
-      [_, n, B, D, w],
+      [h, n, B, D, w],
     ),
     ue = hi({ draggingOver: D.draggingOver, isClone: !1, isDragging: A, mode: A ? D.mode : null }),
     de = (0, M.useMemo)(
       function () {
-        return { draggableId: n, type: f, source: { droppableId: d, index: r } };
+        return { draggableId: n, type: f, source: { droppableId: d, index: i } };
       },
-      [n, d, r, f],
+      [n, d, i, f],
     );
   return M.createElement(
     M.Fragment,
     null,
-    N ? null : t(le, ue, de),
+    ee ? null : t(le, ue, de),
     oe && M.createElement(si, { ref: z }),
   );
 }
@@ -3327,29 +3327,29 @@ function yi(e) {
     i = e.draggableId,
     a = e.index,
     o = e.draggingOver,
-    c = e.style,
-    l = e.getContainerForClone,
-    u = l === void 0 ? vi : l,
-    d = e.mode,
-    f = P().contextId,
-    p = (0, M.useCallback)(
+    s = e.style,
+    c = e.getContainerForClone,
+    l = c === void 0 ? vi : c,
+    u = e.mode,
+    d = N().contextId,
+    f = (0, M.useCallback)(
       function (e) {
-        e && ce({ contextId: f, draggableId: i })?.focus();
+        e && ce({ contextId: d, draggableId: i })?.focus();
       },
-      [f, i],
+      [d, i],
     );
   return (0, X.createPortal)(
     t(
       (0, M.useMemo)(
         function () {
           return {
-            innerRef: p,
-            draggableProps: s(s(s({}, I.draggable.contextId, f), I.draggable.id, i), `style`, c),
-            dragHandleProps: s(
-              s(
-                s(
-                  s(
-                    s({ role: `button`, "aria-describedby": Zt(f) }, I.dragHandle.contextId, f),
+            innerRef: f,
+            draggableProps: C(C(C({}, I.draggable.contextId, d), I.draggable.id, i), `style`, s),
+            dragHandleProps: C(
+              C(
+                C(
+                  C(
+                    C({ role: `button`, "aria-describedby": Zt(d) }, I.dragHandle.contextId, d),
                     I.dragHandle.draggableId,
                     i,
                   ),
@@ -3364,9 +3364,9 @@ function yi(e) {
             ),
           };
         },
-        [f, i, p, c],
+        [d, i, f, s],
       ),
-      hi({ draggingOver: o, isClone: !0, isDragging: !0, mode: d }),
+      hi({ draggingOver: o, isClone: !0, isDragging: !0, mode: u }),
       (0, M.useMemo)(
         function () {
           return { draggableId: i, type: r, source: { droppableId: n, index: a } };
@@ -3374,38 +3374,38 @@ function yi(e) {
         [i, n, a, r],
       ),
     ),
-    u(),
+    l(),
   );
 }
 function bi(e) {
   var t = e.children,
     n = e.droppableId,
-    r = e.type,
-    i = e.getContainerForClone,
-    a = P(),
-    o = a.contextId,
-    s = a.getDragState,
-    c = te(),
-    l = g((0, M.useReducer)(mi, li), 2),
-    u = l[0],
-    d = l[1],
-    f = br();
+    i = e.type,
+    a = e.getContainerForClone,
+    o = N(),
+    s = o.contextId,
+    c = o.getDragState,
+    l = F(),
+    u = r((0, M.useReducer)(mi, li), 2),
+    d = u[0],
+    f = u[1],
+    p = br();
   if (
     ((0, M.useEffect)(
       function () {
         return U(
-          f({
+          p({
             onPendingDragStart: function (e) {
               var t = e.start,
                 r = e.droppable;
               if (n === t.source.droppableId) {
                 if (t.mode === `FLUID`)
-                  return d({ type: `START_POINTER_DRAG`, payload: { start: t } });
+                  return f({ type: `START_POINTER_DRAG`, payload: { start: t } });
                 if (t.mode === `SNAP`) {
-                  var i = s();
+                  var i = c();
                   return (
                     j(i.isDragging && i.draggableDimensions),
-                    d({
+                    f({
                       type: `START_KEYBOARD_DRAG`,
                       payload: {
                         start: t,
@@ -3420,20 +3420,20 @@ function bi(e) {
             onPendingDragUpdate: function (e) {
               var t = e.update,
                 n = e.droppable;
-              u.type === `dragging` &&
-                u.draggableId === t.draggableId &&
-                (d({ type: `UPDATE_DRAG`, payload: { update: t } }),
+              d.type === `dragging` &&
+                d.draggableId === t.draggableId &&
+                (f({ type: `UPDATE_DRAG`, payload: { update: t } }),
                 t.mode === `SNAP` &&
                   queueMicrotask(function () {
-                    if (s().isDragging) {
-                      var e = qr(o),
+                    if (c().isDragging) {
+                      var e = qr(s),
                         r = e ? e.getBoundingClientRect() : null,
                         i = Kr();
-                      d({
+                      f({
                         type: `UPDATE_KEYBOARD_PREVIEW`,
                         payload: {
                           update: t,
-                          draggableDimensions: c,
+                          draggableDimensions: l,
                           droppable: n,
                           placeholderRect: r,
                           dropIndicatorRect: i ? i.getBoundingClientRect() : null,
@@ -3444,38 +3444,38 @@ function bi(e) {
             },
             onBeforeDragEnd: function (e) {
               var t = e.draggableId;
-              u.type === `dragging` && t === u.draggableId && d({ type: `DROP` });
+              d.type === `dragging` && t === d.draggableId && f({ type: `DROP` });
             },
           }),
           Ht({
             canMonitor: function (e) {
               var t = e.source;
-              return V(t.data) ? t.data.contextId === o && t.data.droppableId === n : !1;
+              return V(t.data) ? t.data.contextId === s && t.data.droppableId === n : !1;
             },
             onDrag: function (e) {
               var t = e.location;
-              d({ type: `UPDATE_POINTER_PREVIEW`, payload: { pointerLocation: t } });
+              f({ type: `UPDATE_POINTER_PREVIEW`, payload: { pointerLocation: t } });
             },
           }),
         );
       },
-      [n, o, f, u, c, s],
+      [n, s, p, d, l, c],
     ),
-    u.type !== `dragging`)
+    d.type !== `dragging`)
   )
     return null;
-  var p = ti({ draggableDimensions: c, draggableState: u });
+  var m = ti({ draggableDimensions: l, draggableState: d });
   return M.createElement(
     yi,
     {
       droppableId: n,
-      type: r,
-      draggableId: u.draggableId,
-      index: u.start.index,
-      draggingOver: u.draggingOver,
-      mode: u.mode,
-      style: p,
-      getContainerForClone: i,
+      type: i,
+      draggableId: d.draggableId,
+      index: d.start.index,
+      draggingOver: d.draggingOver,
+      mode: d.mode,
+      style: m,
+      getContainerForClone: a,
     },
     t,
   );
@@ -3587,16 +3587,16 @@ function ji(e) {
   return c ? Di({ element: c, isForwardEdge: !0, mode: i, direction: r, contextId: a }) : null;
 }
 var Mi = 4,
-  Ni = v({
+  Ni = h({
     background: `var(--ds-border-brand, #1868DB)`,
     scrollMarginTop: Mi,
     scrollMarginBottom: Mi + 1,
   }),
-  Pi = v({ position: `absolute`, top: 0, left: 0 }),
-  Fi = v({ opacity: 0 }),
+  Pi = h({ position: `absolute`, top: 0, left: 0 }),
+  Fi = h({ opacity: 0 }),
   Ii = {
-    horizontal: v({ width: 2, height: `100%`, marginLeft: -2 }),
-    vertical: v({ width: `100%`, height: 2, marginTop: -2 }),
+    horizontal: h({ width: 2, height: `100%`, marginLeft: -2 }),
+    vertical: h({ width: `100%`, height: 2, marginTop: -2 }),
   };
 function Li(e) {
   var t = e.direction,
@@ -3606,8 +3606,8 @@ function Li(e) {
   var i = Q[t],
     a = i.mainAxis,
     o = i.crossAxis;
-  return s(
-    s(
+  return C(
+    C(
       { transform: `${a.style.transform}(${n.mainAxis.offset - r}px)` },
       o.style.length,
       n.crossAxis.length,
@@ -3616,34 +3616,34 @@ function Li(e) {
     n.crossAxis.offset,
   );
 }
-var Ri = s({}, L.dropIndicator, ``),
+var Ri = C({}, L.dropIndicator, ``),
   zi = function (e) {
     var t = e.direction,
       n = e.mode,
-      r = P(),
-      i = r.contextId,
-      o = r.getDragState,
-      s = (0, M.useRef)(null),
-      c = g((0, M.useState)(null), 2),
-      l = c[0],
-      u = c[1],
-      d = g((0, M.useState)(!1), 2),
-      f = d[0],
-      p = d[1],
-      m = br(),
-      h = (0, M.useCallback)(
+      i = N(),
+      a = i.contextId,
+      o = i.getDragState,
+      c = (0, M.useRef)(null),
+      l = r((0, M.useState)(null), 2),
+      d = l[0],
+      f = l[1],
+      p = r((0, M.useState)(!1), 2),
+      m = p[0],
+      h = p[1],
+      g = br(),
+      _ = (0, M.useCallback)(
         function (e) {
           var r = e.targetLocation,
-            a = e.source,
+            i = e.source,
             o = e.destination;
-          if (!r) return u(null);
-          var s = H(a, o);
+          if (!r) return f(null);
+          var s = H(i, o);
           return (
-            p(s),
-            u(ji({ targetLocation: r, isInHomeLocation: s, direction: t, mode: n, contextId: i }))
+            h(s),
+            f(ji({ targetLocation: r, isInHomeLocation: s, direction: t, mode: n, contextId: a }))
           );
         },
-        [i, t, n],
+        [a, t, n],
       );
     ((0, M.useLayoutEffect)(
       function () {
@@ -3652,38 +3652,38 @@ var Ri = s({}, L.dropIndicator, ``),
           var t = e.targetLocation,
             n = e.sourceLocation;
           return (
-            h({ targetLocation: t, destination: Xt({ start: n, target: t }), source: n }),
-            m({
+            _({ targetLocation: t, destination: Xt({ start: n, target: t }), source: n }),
+            g({
               onPrePendingDragUpdate: function (e) {
                 var t = e.update,
                   n = e.targetLocation,
                   r = t.destination,
                   i = r === void 0 ? null : r,
                   a = t.source;
-                h({ targetLocation: n, source: a, destination: i });
+                _({ targetLocation: n, source: a, destination: i });
               },
             })
           );
         }
       },
-      [i, t, o, n, m, h],
+      [a, t, o, n, g, _],
     ),
       (0, M.useLayoutEffect)(
         function () {
-          if (l !== null) {
+          if (d !== null) {
             var e = o();
             if (!(!e.isDragging || e.mode !== `SNAP`)) {
-              var t = s.current;
+              var t = c.current;
               (j(t instanceof HTMLElement), t.scrollIntoView({ block: `nearest` }));
             }
           }
         },
-        [l, o],
+        [d, o],
       ));
-    var _ = Q[t].mainAxis,
-      v = Li({ direction: t, dimensions: l, indicatorOffset: s.current ? s.current[_.offset] : 0 }),
-      y = n === `virtual`;
-    return b(`div`, a({ ref: s, css: [Ni, Ii[t], y && Pi, f && Fi], style: v }, Ri));
+    var v = Q[t].mainAxis,
+      y = Li({ direction: t, dimensions: d, indicatorOffset: c.current ? c.current[v.offset] : 0 }),
+      b = n === `virtual`;
+    return u(`div`, s({ ref: c, css: [Ni, Ii[t], b && Pi, m && Fi], style: y }, Ri));
   };
 function Bi(e, t) {
   var n = Object.keys(e);
@@ -3702,7 +3702,7 @@ function Vi(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? Bi(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -3747,7 +3747,7 @@ function Wi(e) {
     i = e.direction,
     a = e.isDropDisabled,
     o = (0, M.useRef)(null),
-    s = P(),
+    s = N(),
     c = s.contextId,
     l = s.getDragState,
     u = l();
@@ -3787,72 +3787,72 @@ function Wi(e) {
 function Gi(e) {
   var t = e.children,
     n = e.droppableId,
-    r = e.type,
-    i = r === void 0 ? `DEFAULT` : r,
-    a = e.direction,
-    o = a === void 0 ? `vertical` : a,
+    i = e.type,
+    a = i === void 0 ? `DEFAULT` : i,
+    o = e.direction,
+    s = o === void 0 ? `vertical` : o,
     c = e.mode,
     l = c === void 0 ? `standard` : c,
     u = e.renderClone,
     d = e.getContainerForClone,
     f = e.isDropDisabled,
-    m = f === void 0 ? !1 : f,
-    h = Jr(m),
-    _ = P(),
-    v = _.contextId,
-    y = _.droppableRegistry,
-    b = Se({ contextId: v, droppableId: n, getIsDropDisabled: h }),
-    x = (0, M.useRef)(null),
-    S = (0, M.useCallback)(
+    p = f === void 0 ? !1 : f,
+    m = Jr(p),
+    h = N(),
+    g = h.contextId,
+    _ = h.droppableRegistry,
+    v = Se({ contextId: g, droppableId: n, getIsDropDisabled: m }),
+    b = (0, M.useRef)(null),
+    x = (0, M.useCallback)(
       function (e) {
         (e &&
           ae(
             e,
-            s(
-              s(s(s({}, L.droppable.type, i), L.droppable.direction, o), I.droppable.id, n),
+            C(
+              C(C(C({}, L.droppable.type, a), L.droppable.direction, s), I.droppable.id, n),
               I.droppable.contextId,
-              v,
+              g,
             ),
           ),
-          (x.current = e));
+          (b.current = e));
       },
-      [v, o, n, i],
+      [g, s, n, a],
     ),
-    C = g((0, M.useReducer)(Ui, Hi), 2),
-    w = C[0],
-    T = C[1],
+    S = r((0, M.useReducer)(Ui, Hi), 2),
+    w = S[0],
+    T = S[1],
     E = w.draggingFromThisWith,
     D = w.draggingOverWith,
     O = w.isDraggingOver,
     k = Ur();
   (0, M.useEffect)(
     function () {
-      var e = x.current;
+      var e = b.current;
       return (
-        p(e instanceof HTMLElement, "innerRef must provide an `HTMLElement`"),
+        y(e instanceof HTMLElement, "innerRef must provide an `HTMLElement`"),
         U(
-          y.register({
+          _.register({
             droppableId: n,
-            type: i,
-            isDropDisabled: m,
+            type: a,
+            isDropDisabled: p,
             parentDroppableId: k,
             element: e,
-            direction: o,
+            direction: s,
             mode: l,
           }),
           Vt({
             element: e,
             getData: function (t) {
               var n = t.input;
-              return ge(b, {
+              return ge(v, {
                 element: e,
                 input: n,
-                allowedEdges: o === `vertical` ? [`top`, `bottom`] : [`left`, `right`],
+                allowedEdges: s === `vertical` ? [`top`, `bottom`] : [`left`, `right`],
               });
             },
             canDrop: function (e) {
               var t = e.source;
-              return !V(t.data) || m ? !1 : t.data.contextId === v && t.data.type === i;
+              return !V(t.data) || p ? !1 : t.data.contextId === g && t.data.type === a;
             },
             onDragLeave: function () {
               T({ type: `DRAG_CLEAR` });
@@ -3861,13 +3861,13 @@ function Gi(e) {
         )
       );
     },
-    [b, n, v, m, i, y, k, o, l],
+    [v, n, g, p, a, _, k, s, l],
   );
   var A = br();
   (0, M.useEffect)(
     function () {
       function e(e) {
-        var t = e.type === i,
+        var t = e.type === a,
           r = e.destination?.droppableId === n;
         return t && ((!O && r) || (O && !r));
       }
@@ -3886,25 +3886,25 @@ function Gi(e) {
         },
       });
     },
-    [n, O, A, i],
+    [n, O, A, a],
   );
   var j = (0, M.useMemo)(
       function () {
-        return O ? M.createElement(zi, { direction: o, mode: l }) : null;
+        return O ? M.createElement(zi, { direction: s, mode: l }) : null;
       },
-      [o, O, l],
+      [s, O, l],
     ),
-    N = (0, M.useMemo)(
+    ee = (0, M.useMemo)(
       function () {
         return {
-          innerRef: S,
-          droppableProps: s(s({}, I.droppable.contextId, v), I.droppable.id, n),
+          innerRef: x,
+          droppableProps: C(C({}, I.droppable.contextId, g), I.droppable.id, n),
           placeholder: l === `standard` ? j : null,
         };
       },
-      [v, j, n, l, S],
+      [g, j, n, l, x],
     ),
-    ee = (0, M.useMemo)(
+    te = (0, M.useMemo)(
       function () {
         return {
           draggingFromThisWith: E,
@@ -3915,67 +3915,67 @@ function Gi(e) {
       },
       [E, D, O],
     ),
-    F = x.current,
-    te = O && l === `virtual` && F;
+    P = b.current,
+    F = O && l === `virtual` && P;
   rr(
     function () {
-      if (te && window.getComputedStyle(F).position === `static`) {
-        var e = F.style.position;
+      if (F && window.getComputedStyle(P).position === `static`) {
+        var e = P.style.position;
         return (
-          (F.style.position = `relative`),
+          (P.style.position = `relative`),
           function () {
-            F.style.position = e;
+            P.style.position = e;
           }
         );
       }
     },
-    [F, te],
+    [P, F],
   );
   var ne = !!u,
     re = (0, M.useMemo)(
       function () {
         return {
-          direction: o,
+          direction: s,
           droppableId: n,
           shouldRenderCloneWhileDragging: ne,
-          isDropDisabled: m,
-          type: i,
+          isDropDisabled: p,
+          type: a,
           mode: l,
         };
       },
-      [o, n, ne, m, i, l],
+      [s, n, ne, p, a, l],
     ),
-    ie = E && l === `virtual` && F;
+    ie = E && l === `virtual` && P;
   return M.createElement(
     Vr,
     { value: re },
-    t(N, ee),
-    te && (0, X.createPortal)(j, F),
+    t(ee, te),
+    F && (0, X.createPortal)(j, P),
     ie &&
       (0, X.createPortal)(
         M.createElement(Wi, {
           droppableId: n,
           draggableId: E,
-          type: i,
-          direction: o,
-          isDropDisabled: m,
+          type: a,
+          direction: s,
+          isDropDisabled: p,
         }),
-        F,
+        P,
       ),
-    u && M.createElement(bi, { droppableId: n, type: i, getContainerForClone: d }, u),
+    u && M.createElement(bi, { droppableId: n, type: a, getContainerForClone: d }, u),
   );
 }
-i();
+l();
 var Ki = [`isRanking`, `isRankingItem`, `testId`],
   qi = (0, M.forwardRef)(function (e, t) {
     var n = e.isRanking,
       r = e.isRankingItem,
       i = e.testId,
-      s = c(e, Ki);
+      a = w(e, Ki);
     return M.createElement(
-      y,
-      a({ ref: t, testId: i }, s, {
-        className: o([
+      _,
+      s({ ref: t, testId: i }, a, {
+        className: S([
           n && `_1e0c1ule`,
           r && `_2rkolb4i _bfhk1bhr _16qs130s`,
           `_12y3e4h9 _mizu1v1w _ra3xnqa1`,
@@ -3983,17 +3983,17 @@ var Ki = [`isRanking`, `isRankingItem`, `testId`],
       }),
     );
   });
-i();
+l();
 var Ji = [`isRanking`, `innerRef`],
   Yi = function (e) {
     var t = e.isRanking,
       n = e.innerRef,
-      r = c(e, Ji);
-    return M.createElement(h, a({}, r, { innerRef: n, className: o([t && `_vchhusvi`]) }));
+      r = w(e, Ji);
+    return M.createElement(m, s({}, r, { innerRef: n, className: S([t && `_vchhusvi`]) }));
   },
   Xi = [`content`, `testId`];
 function Zi(e, t, n) {
-  return ((t = T(t)), d(e, Qi() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
+  return ((t = a(t)), i(e, Qi() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
 }
 function Qi() {
   try {
@@ -4003,14 +4003,14 @@ function Qi() {
     return !!e;
   })();
 }
-var $i = _(
+var $i = b(
     (function (e) {
       function t() {
-        return (C(this, t), Zi(this, t, arguments));
+        return (g(this, t), Zi(this, t, arguments));
       }
       return (
-        S(t, e),
-        r(t, [
+        c(t, e),
+        o(t, [
           {
             key: `render`,
             value: function () {
@@ -4019,31 +4019,31 @@ var $i = _(
                 n = e.head,
                 r = e.isFixedSize,
                 i = e.isRanking,
-                o = e.refWidth,
-                s = e.innerRef,
-                l = e.testId,
-                u = t.content,
-                d = t.testId,
-                f = c(t, Xi),
-                p = n || {},
-                m = p.shouldTruncate,
-                h = p.width,
-                g = x(i, o);
+                a = e.refWidth,
+                o = e.innerRef,
+                c = e.testId,
+                l = t.content,
+                u = t.testId,
+                d = w(t, Xi),
+                f = n || {},
+                p = f.shouldTruncate,
+                m = f.width,
+                h = v(i, a);
               return M.createElement(
                 Yi,
-                a({}, f, {
+                s({}, d, {
                   isFixedSize: r,
-                  shouldTruncate: m,
-                  width: h,
+                  shouldTruncate: p,
+                  width: m,
                   isRanking: i,
-                  style: g,
+                  style: h,
                   onKeyDown: function (e) {
                     return e.stopPropagation();
                   },
-                  innerRef: s,
-                  "data-testid": d || (l && `${l}--rankable--table--body--cell`),
+                  innerRef: o,
+                  "data-testid": u || (c && `${c}--rankable--table--body--cell`),
                 }),
-                u,
+                l,
               );
             },
           },
@@ -4069,7 +4069,7 @@ function na(e) {
     var n = arguments[t] == null ? {} : arguments[t];
     t % 2
       ? ta(Object(n), !0).forEach(function (t) {
-          s(e, t, n[t]);
+          C(e, t, n[t]);
         })
       : Object.getOwnPropertyDescriptors
         ? Object.defineProperties(e, Object.getOwnPropertyDescriptors(n))
@@ -4080,7 +4080,7 @@ function na(e) {
   return e;
 }
 function ra(e, t, n) {
-  return ((t = T(t)), d(e, ia() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
+  return ((t = a(t)), i(e, ia() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
 }
 function ia() {
   try {
@@ -4090,15 +4090,15 @@ function ia() {
     return !!e;
   })();
 }
-var aa = _(
+var aa = b(
   (function (e) {
     function t() {
       var e;
-      C(this, t);
+      g(this, t);
       var n = [...arguments];
       return (
         (e = ra(this, t, [].concat(n))),
-        s(e, `innerRef`, function (t) {
+        C(e, `innerRef`, function (t) {
           return function (n) {
             (t(n), typeof e.props.innerRef == `function` && e.props.innerRef(n));
           };
@@ -4107,8 +4107,8 @@ var aa = _(
       );
     }
     return (
-      S(t, e),
-      r(t, [
+      c(t, e),
+      o(t, [
         {
           key: `render`,
           value: function () {
@@ -4117,48 +4117,48 @@ var aa = _(
               n = t.row,
               r = t.head,
               i = t.isFixedSize,
-              o = t.isRanking,
-              s = t.refWidth,
-              l = t.rowIndex,
-              u = t.isRankingDisabled,
-              d = t.isHighlighted,
-              f = t.testId,
-              p = n.cells,
-              m = n.testId,
-              h = n.key,
-              g = n.isHighlighted,
-              _ = c(n, ea),
-              v = x(o, s);
-            if (typeof h != `string` && !u)
+              a = t.isRanking,
+              o = t.refWidth,
+              c = t.rowIndex,
+              l = t.isRankingDisabled,
+              u = t.isHighlighted,
+              d = t.testId,
+              f = n.cells,
+              p = n.testId,
+              m = n.key,
+              h = n.isHighlighted,
+              g = w(n, ea),
+              _ = v(a, o);
+            if (typeof m != `string` && !l)
               throw Error(
                 `dynamic-table: ranking is not possible because table row does not have a key. Add the key to the row or disable ranking.`,
               );
             return M.createElement(
               _i,
-              { draggableId: h || l.toString(), index: l, isDragDisabled: u },
+              { draggableId: m || c.toString(), index: c, isDragDisabled: l },
               function (t, n) {
                 return M.createElement(
                   qi,
-                  a({}, _, t.dragHandleProps, t.draggableProps, {
+                  s({}, g, t.dragHandleProps, t.draggableProps, {
                     role: void 0,
                     "aria-labelledby": void 0,
                     "aria-describedby": t.dragHandleProps?.[`aria-describedby`],
                     ref: e.innerRef(t.innerRef),
-                    style: na(na({}, t.draggableProps.style), v),
-                    isHighlighted: d || g,
-                    isRanking: o,
+                    style: na(na({}, t.draggableProps.style), _),
+                    isHighlighted: u || h,
+                    isRanking: a,
                     isRankingItem: n.isDragging,
-                    testId: m || (f && `${f}--rankable--table--body--row`),
+                    testId: p || (d && `${d}--rankable--table--body--row`),
                   }),
-                  p.map(function (e, t) {
+                  f.map(function (e, t) {
                     var n = (r || { cells: [] }).cells[t];
                     return M.createElement($i, {
                       head: n,
                       cell: e,
-                      isRanking: o,
+                      isRanking: a,
                       key: e.key || t,
                       isFixedSize: i,
-                      testId: f,
+                      testId: d,
                     });
                   }),
                 );
@@ -4171,7 +4171,7 @@ var aa = _(
   })(M.Component),
 );
 function oa(e, t, n) {
-  return ((t = T(t)), d(e, sa() ? Reflect.construct(t, n || [], T(e).constructor) : t.apply(e, n)));
+  return ((t = a(t)), i(e, sa() ? Reflect.construct(t, n || [], a(e).constructor) : t.apply(e, n)));
 }
 function sa() {
   try {
@@ -4195,16 +4195,16 @@ var ca = function (e, t) {
   la = (function (e) {
     function t() {
       var e;
-      C(this, t);
+      g(this, t);
       var n = [...arguments];
       return (
         (e = oa(this, t, [].concat(n))),
-        s(e, `onBeforeDragStart`, function (t) {
+        C(e, `onBeforeDragStart`, function (t) {
           var n = t.draggableId,
             r = { index: t.source.index, key: n };
           e.props.onRankStart(r);
         }),
-        s(e, `onDragEnd`, function (t) {
+        C(e, `onDragEnd`, function (t) {
           var n = e.props,
             r = n.pageRows,
             i = n.onRankEnd,
@@ -4215,8 +4215,8 @@ var ca = function (e, t) {
       );
     }
     return (
-      S(t, e),
-      r(t, [
+      c(t, e),
+      o(t, [
         {
           key: `render`,
           value: function () {
@@ -4225,8 +4225,8 @@ var ca = function (e, t) {
               n = e.pageRows,
               r = e.head,
               i = e.isFixedSize,
-              o = e.isRanking,
-              s = e.isRankingDisabled,
+              a = e.isRanking,
+              o = e.isRankingDisabled,
               c = e.testId,
               l = e.forwardedRef;
             return M.createElement(
@@ -4234,11 +4234,11 @@ var ca = function (e, t) {
               { onBeforeDragStart: this.onBeforeDragStart, onDragEnd: this.onDragEnd },
               M.createElement(
                 Gi,
-                { droppableId: `dynamic-table-droppable`, isDropDisabled: s },
+                { droppableId: `dynamic-table-droppable`, isDropDisabled: o },
                 function (e) {
                   return M.createElement(
                     `tbody`,
-                    a(
+                    s(
                       {
                         "data-testid": c,
                         ref: function (t) {
@@ -4251,12 +4251,12 @@ var ca = function (e, t) {
                     n.map(function (e, n) {
                       return M.createElement(aa, {
                         head: r,
-                        isRanking: o,
+                        isRanking: a,
                         isFixedSize: i,
                         key: e.key,
                         rowIndex: n,
                         row: e,
-                        isRankingDisabled: s,
+                        isRankingDisabled: o,
                         isHighlighted:
                           t !== void 0 && (typeof t == `number` ? t === n : t.indexOf(n) > -1),
                         testId: c && `${c}--${e.key}--rankable--table--row`,
@@ -4272,9 +4272,9 @@ var ca = function (e, t) {
       ])
     );
   })(M.Component),
-  ua = u(
+  ua = p(
     M.forwardRef(function (e, t) {
-      return M.createElement(la, a({}, e, { forwardedRef: t }));
+      return M.createElement(la, s({}, e, { forwardedRef: t }));
     }),
   );
 export { ua as default };
