+++ examples/forge-sql-orm-example-optimistic-locking/static/forge-orm-example/build-static/assets/custom-theme-WUDk5RIJ.js
@@ -10,7 +10,7 @@ import {
   s as c,
   u as l,
   w as u,
-} from "./index-CRQqC8L5.js";
+} from "./index-RNH_WMbf.js";
 var d = {
   "color.text.brand": `#669DF1`,
   "elevation.surface.overlay": `#2B2C2F`,
