+++ examples/forge-sql-orm-example-optimistic-locking/static/forge-orm-example/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "dependencies": {
-    "@atlaskit/button": "^23.11.4",
+    "@atlaskit/button": "^23.11.5",
     "@atlaskit/css-reset": "^7.4.5",
     "@atlaskit/dynamic-table": "^18.4.1",
     "@atlaskit/tabs": "^19.1.0",
