+++ examples/forge-sql-orm-example-vector/static/forge-orm-example/build-static/assets/custom-theme-BT2SGvsq.js
@@ -10,7 +10,7 @@ import {
   s as c,
   t as l,
   u,
-} from "./index-B6mhip_G.js";
+} from "./index-EQxheKR4.js";
 var d = {
   "color.text.brand": `#669DF1`,
   "elevation.surface.overlay": `#2B2C2F`,
