+++ examples/forge-sql-orm-example-vector/static/forge-orm-example/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@atlaskit/button": "^23.11.4",
+    "@atlaskit/button": "^23.11.5",
     "@atlaskit/css-reset": "^7.4.5",
     "@atlaskit/primitives": "^19.0.1",
     "@atlaskit/section-message": "^8.12.14",
